AE Assessment - ADaM Mapping

Convert from ADaM format to input format for Safety Assessment.

AE_Map_Adam(dfADSL, dfADAE, mapping = NULL, bQuiet = TRUE)

Arguments

dfADSL

ADaM demographics data with the following required columns: USUBJID, SITEID, TRTEDT (end date), TRTSDT (start date)

dfADAE

ADaM AE data with the following required columns: USUBJID

mapping

List containing expected columns in each data set. By default, mapping for dfAE is: strIDCol = "SUBJID". By default, mapping for dfRDSL is: strIDCol = "SubjectID", strSiteCol = "SiteID", and strExposureCol = "TimeOnTreatment". TODO: add more descriptive info or reference to mapping.

bQuiet

Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.

Value

Data frame with one record per person data frame with columns: SubjectID, SiteID, Count (Number of Adverse Events), Exposure (Time on Treatment in Days), Rate (AEs/Day)

Details

This function maps from ADaM ADSL and ADAE data to the required input for AE_Assess.

Data Specification

This function creates an input dataset for the Adverse Event Assessment (AE_Assess) by adding Adverse Event Counts (from dfADAE) to basic subject-level data (from dfADSL).

The following columns are required:

dfADSL

USUBJID - Unique subject ID

SITEID - Site ID

TRTEDT - Treatment End date

TRTSDT - Treatment Start date

dfADAE

USUBJID - Unique subject ID

Note that the function can generate data summaries for specific types of AEs by passing filtered ADAE data to dfADAE.

Examples

dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )