Chi-squared Test Analysis

Creates Analysis results data for count data using the chi-squared test

Analyze_Chisq(dfTransformed, strOutcome = "TotalCount")

Arguments

dfTransformed

data.frame in format produced by Transform_EventCount

strOutcome

required, name of column in dfTransformed dataset to perform the chi-squared test on. Default is "TotalCount".

Value

data.frame with one row per site, columns: SiteID, TotalCount, TotalCount_Other, N, N_Other, Prop, Prop_Other, Statistic, PValue

Details

Analyzes count data using the chi-squared test

Statistical Methods

TODO Coming soon ...

Data Specification

The input data (dfTransformed) for Analyze_Chisq is typically created using Transform_EventCount and should be one record per site with required columns for:

SiteID - Site ID

N - Total number of participants at site

TotalCount - Total number of participants at site with event of interest

Examples

dfInput <- Disp_Map(dfDisp = safetyData::adam_adsl, strCol = "DCREASCD",strReason = "Adverse Event")
dfTransformed <- Transform_EventCount( dfInput, strCountCol = 'Count' )
dfAnalyzed <- Analyze_Chisq( dfTransformed )
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect