Identity Analysis.

Analyze_Identity(dfTransformed, strValueCol = "Metric", bQuiet = TRUE)

Arguments

dfTransformed

data.frame created by Transform_Count

strValueCol

character Name of column that will be copied as Score

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one row per site with columns: GroupID, TotalCount, Metric, and Score.

Details

Used in the data pipeline between Transform and Flag to rename KRI and Score columns.

More information can be found in The Identity Method of the KRI Method vignette.

Data Specification

The input data (dfTransformed) for Analyze_Identity is typically created using Transform_Rate and should be one record per site with required columns for:

GroupID - Site ID

Numerator - Total number of participants at site with event of interest.

Denominator - Total number of participants at site/Total number of days of exposure at site.

Metric - Proportion of participants at site with event of interest/Rate of events at site (Numerator / Denominator).

Examples

dfInput <- Consent_Map_Raw()
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")
dfAnalyzed <- Analyze_Identity(dfTransformed)