Poisson Analysis - Predicted Boundaries.

Analyze_Poisson_PredictBounds(
  dfTransformed,
  vThreshold = c(-5, 5),
  nStep = NULL
)

Arguments

dfTransformed

data.frame data.frame in format produced by Transform_Rate. Must include GroupID, GroupLevel, N, Numerator and Denominator.

vThreshold

numeric upper and lower boundaries in residual space. Should be identical to the thresholds used AE_Assess().

nStep

numeric step size of imputed bounds.

Value

data.frame containing predicted boundary values with upper and lower bounds across the range of observed values.

Details

Fits a Poisson model to site-level data and then calculates predicted count values and upper- and lower- bounds for across the full range of exposure values.

Statistical Methods

This function fits a Poisson model to site-level data and then calculates residuals for each site. The Poisson model is run using standard methods in the stats package by fitting a glm model with family set to poisson using a "log" link. Upper and lower boundary values are then calculated using the method described here TODO: Add link.

Data Specification

The input data (dfTransformed) for Analyze_Poisson_PredictBounds is typically created using Transform_Rate and should be one record per site with columns for:

GroupID - Unique subject ID

GroupLevel - Group Type

Numerator - Number of Events

Denominator - Number of days of exposure

Examples

dfTransformed <- Transform_Rate(analyticsInput)

dfBounds <- Analyze_Poisson_PredictBounds(dfTransformed, c(-5, 5))
#> → nStep was not provided. Setting default step to 0.00469584690363958