QTL Analysis for Binary and Rate Outcomes.

Analyze_QTL(
  dfTransformed,
  nConfLevel = 0.95,
  strOutcome = "binary",
  bQuiet = TRUE
)

Arguments

dfTransformed

data.frame in format produced by Transform_Rate

nConfLevel

numeric specified confidence interval for QTL analysis.

strOutcome

character indicates statistical test used for QTL analysis. One of rate or binary.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one row with columns: GroupID, Numerator, Denominator, Metric, Method, ConfLevel, Estimate, LowCI, UpCI and Score.

Details

Creates confidence intervals for the observed proportion of participants for the event of interest. Uses the exact binomial (stats::binom.test) or poisson test (stats::poisson.test).

Statistical Methods

TODO Coming soon ...

Data Specification

The input data (dfTransformed) for Analyze_QTL is typically created using Transform_Rate and should be one record for the entire study with required columns for:

GroupID - GroupID should be the StudyID

N - Total number of participants at site

Numerator - Total number of participants at site with event of interest

Denominator - Total number of participants at a site

Metric - Proportion of participants at site with event of interest

Examples

dfInput <- Disp_Map_Raw()
dfTransformed <- Transform_Rate(dfInput,
  strNumeratorCol = "Count",
  strDenominatorCol = "Total",
  strGroupCol = "StudyID"
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "binary")
dfFlagged <- Flag_QTL(dfAnalyzed, vThreshold = 0.2)

dfInput <- PD_Map_Raw_Binary()
dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "StudyID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Total"
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "rate")
dfFlagged <- Flag_QTL(dfAnalyzed, vThreshold = 0.01)