Consent Assessment

Consent Assessment

Consent_Assess(
  dfInput,
  nThreshold = 0.5,
  lTags = list(Assessment = "Consent"),
  bChart = TRUE
)

Arguments

dfInput

input data with one record per person and the following required columns: SubjectID, SiteID, Count.

nThreshold

Any sites where 'N' is greater than nThreshold will be flagged. Default value is 0.5, which flags any site with one or more subjects meeting any of the criteria.

lTags

named list of tags describing the assessment. lTags is returned as part of the assessment (lAssess$lTags) and each tag is added as columns in lassess$dfSummary. Default is list(Assessment="Consent")

bChart

should visualization be created? TRUE (default) or FALSE.

Value

A list containing all data and metadata in the standard data pipeline (dfInput, dfTransformed, dfAnalyzed, dfFlagged, dfSummary, strFunctionName, lParams and lTags) is returned.

Details

The Consent Assessment uses the standard GSM data pipeline (TODO add link to data vignette) to flag sites with participants who started study activities before consent was finalized. More details regarding the data pipeline and statistical methods are described below.

Data Specification

The input data (dfInput) for Consent Assessment is typically created using Consent_Map_Raw and should be one record per person with columns for:

SubjectID - Unique subject ID

SiteID - Site ID

Count - Number of findings of errors/outliers.

The Assessment

Transform_EventCount creates dfTransformed.

Flag creates dfFlagged.

Summarize creates dfSummary.

Statistical Assumptions

This Assessment finds any sites where one or more subjects meets any of the following criteria: No Consent, Missing Consent, Missing Randomization Date, or Consent date later in time than the Randomization Date. 'N' in the summary represents the number of subjects in a study that meet one or more criteria. Sites With N greater than user specified nThreshold will be flagged.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
raw_consent <- clindata::raw_ic_elig %>%
select(SUBJID, CONSDAT = DSSTDAT_RAW) %>%
 mutate(
     CONSCAT_STD = "MAINCONSENT",
     CONSYN = "Y",
     CONSDAT = as.Date(CONSDAT, format = "%d %B %Y")
   ) %>%
   filter(SUBJID != "")

dfInput <- Consent_Map_Raw(
   dfCONSENT = clindata::rawplus_consent,
   dfSUBJ = clindata::rawplus_subj,
   strConsentTypeValue = "MAINCONSENT",
   strConsentStatusValue="Y"
)

Consent_Summary <- Consent_Assess(dfInput)$dfSummary