Consolidate stratified assessment outputs.

Consolidates multiple stratified assessment outputs by stacking the data frames returned by an *_Assess function and generating a paneled data visualization.

ConsolidateStrata(lOutput, lStratifiedOutput, bQuiet = TRUE)

Arguments

lOutput

list The output from RunWorkflow()

lStratifiedOutput

list Multiple outputs from RunWorkflow()

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list with group, steps, path, name, lData, lChecks, bStatus, lWorkflowChecks, and lResults added based on the results of the execution of assessment$workflow.

Details

Examples

lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae
)
lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lWorkflow <- MakeWorkflowList(bRecursive = TRUE, strNames = "aeGrade")$aeGrade
lOutput <- RunWorkflow(lWorkflow, lData = lData, lMapping = lMapping)

lStratifiedWorkflow <- MakeStratifiedAssessment(
  lWorkflow,
  lData,
  lMapping
)
lStratifiedOutput <- lStratifiedWorkflow %>%
  purrr::map(~ RunWorkflow(
    .x,
    lData,
    lMapping
  ))

lConsolidatedOutput <- ConsolidateStrata(
  lOutput,
  lStratifiedOutput
)