Convert raw data entry data to formatted input data to DataEntry_Assess().

DataEntry_Map_Raw(
  dfs = list(dfSUBJ = clindata::rawplus_dm, dfDATAENT = clindata::edc_data_pages),
  lMapping = gsm::Read_Mapping(c("rawplus", "edc")),
  nMaxDataEntryLag = 10,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)

Arguments

dfs

list Input data frame:

dfSUBJ: data.frame Subject-level data with one record per participant.

dfDATAENT: data.frame Data-Point-level data with one record per data entry.

lMapping

list Column metadata with structure domain$key, where key contains the name of the column.

nMaxDataEntryLag

numeric Expected number of days allowed to enter data

bReturnChecks

logical Return input checks from is_mapping_valid()? Default: FALSE

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one record per subject, the input to DataEntry_Assess(). If bReturnChecks is TRUE

DataEntry_Map_Raw returns a named list with:

df: the data frame described above

lChecks: a named list of check results

Details

Data Entry Lag Assessment - Raw Mapping

DataEntry_Map_Raw creates an input dataset for the Data Entry Lag Assessment DataEntry_Assess() by adding Data Entry Lag Counts (derived from dfDATAENT) to basic subject-level data (from dfSUBJ).

Data specification

Domain	Column Key	Default Value	Required?	Require Unique Values?

dfSUBJ	strSiteCol	siteid	TRUE	FALSE

dfSUBJ	strEDCIDCol	subject_nsv	TRUE	TRUE

dfDATACHG	strIDCol	subjectname	TRUE	FALSE

dfDATACHG	strNChangesCol	n_changes	TRUE	FALSE

Examples

# Run with defaults.
dfInput <- DataEntry_Map_Raw()

# Run with error checking and message log.
dfInput <- DataEntry_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)
#> 
#> ── Checking Input Data for `DataEntry_Map_Raw()` ──
#> 
#> ✔ No issues found for `DataEntry_Map_Raw()`
#> 
#> ── Initializing `DataEntry_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ✔ `DataEntry_Map_Raw()` returned output with 1301 rows.