Export gsm analysis script using Data, Mapping, and Workflow.

ExportCode(
  lData,
  lMapping,
  lAssessments,
  bInsertText = FALSE,
  strPath = NULL,
  strFileName = NULL
)

Arguments

lData

list rawplus data to use as inputs.

lMapping

list Standard mapping provided for FilterDomain() and *_Map_Raw() functions.

lAssessments

list The result of running MakeWorkflowList(), or a custom workflow.

bInsertText

logical Should code be inserted into a new .R file? Default: TRUE.

strPath

character Path for where the code should be saved.

strFileName

character Name of file to save.

Examples

if (FALSE) {

lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::ctms_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie,
  dfLB = clindata::rawplus_lb,
  dfSTUDCOMP = clindata::rawplus_studcomp,
  dfSDRGCOMP = clindata::rawplus_sdrgcomp
)

lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lAssessments <- gsm::MakeWorkflowList()

code <- ExportCode(lData, lMapping, lAssessments)
}