Inclusion/Exclusion Assessment

Flag sites exhibiting aberrant or excessive rates of unmet or missing inclusion/exclusion (IE) criteria.

IE_Assess(
  dfInput,
  nThreshold = 0.5,
  lTags = list(Assessment = "IE"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)

Arguments

dfInput

data.frame Input data, a data frame with one record per subject.

nThreshold

numeric Threshold specification. Default: 0.5

lTags

list Assessment tags, a named list of tags describing the assessment that defaults to list(Assessment="IE"). lTags is returned as part of the assessment (lAssess$lTags) and each tag is added as a column in lAssess$dfSummary.

bChart

logical Generate data visualization? Default: TRUE

bReturnChecks

logical Return input checks from is_mapping_valid? Default: FALSE

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list Assessment, a named list with:

each data frame in the data pipeline

dfInput

dfTransformed, returned by Transform_EventCount()

dfAnalyzed, a copy of dfTransformed and input to [gsm::Flag()]

dfFlagged, returned by Flag()

dfSummary, returned by Summarize()

assessment metadata

strFunctionName

lParams

lTags

output(s)

chart

Details

The IE Assessment uses the standard GSM data pipeline to flag sites with IE issues. This assessment detects sites with excessive rates of unmet or missing IE criteria, as defined by nThreshold. The count returned in the summary represents the number of subjects at a given site with at least one unmet or missing IE criterion. Additional details regarding the data pipeline and statistical methods are described below.

Data specification

domain	col_key	col_value	vRequired	vUniqueCols

dfInput	strIDCol	SubjectID	TRUE	TRUE

dfInput	strSiteCol	SiteID	TRUE	FALSE

dfInput	strCountCol	Count	TRUE	FALSE

Examples

dfInput <- IE_Map_Raw()
#> ℹ Intializing merge of domain and subject data
ie_assessment <- IE_Assess(dfInput)