Load assessments from a package/directory

Load assessments from a package/directory

MakeAssessmentList(
  strNames = NULL,
  strPath = "workflow",
  strPackage = "gsm",
  bRecursive = FALSE
)

Arguments

strNames

array of character List of workflows to include. NULL (the default) includes all workflows in the specified locations.

strPath

character The location of assessment YAML files. If package is specified, function will look in /inst folder.

strPackage

character The name of the package with assessments.

bRecursive

logical Find files in nested folders? Default FALSE.

Value

list A list of assessments with workflow and parameter metadata.

Details

MakeAssessmentList() is a utility function that creates a workflow mapping for assessments used in Study_Assess().

Examples

MakeAssessmentList(strPath = "workflow", strPackage = "gsm")
#> $kri0001
#> $kri0001$workflow
#> $kri0001$workflow[[1]]
#> $kri0001$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0001$workflow[[1]]$inputs
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[1]]$output
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[1]]$params
#> $kri0001$workflow[[1]]$params$strDomain
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[1]]$params$strColParam
#> [1] "strTreatmentEmergentCol"
#> 
#> $kri0001$workflow[[1]]$params$strValParam
#> [1] "strTreatmentEmergentVal"
#> 
#> 
#> 
#> $kri0001$workflow[[2]]
#> $kri0001$workflow[[2]]$name
#> [1] "FilterDomain"
#> 
#> $kri0001$workflow[[2]]$inputs
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[2]]$output
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[2]]$params
#> $kri0001$workflow[[2]]$params$strDomain
#> [1] "dfAE"
#> 
#> $kri0001$workflow[[2]]$params$strColParam
#> [1] "strSeriousCol"
#> 
#> $kri0001$workflow[[2]]$params$strValParam
#> [1] "strNonSeriousVal"
#> 
#> 
#> 
#> $kri0001$workflow[[3]]
#> $kri0001$workflow[[3]]$name
#> [1] "AE_Map_Raw"
#> 
#> $kri0001$workflow[[3]]$inputs
#> [1] "dfAE"   "dfSUBJ"
#> 
#> $kri0001$workflow[[3]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0001$workflow[[4]]
#> $kri0001$workflow[[4]]$name
#> [1] "AE_Assess"
#> 
#> $kri0001$workflow[[4]]$inputs
#> [1] "dfInput"
#> 
#> $kri0001$workflow[[4]]$output
#> [1] "lResults"
#> 
#> $kri0001$workflow[[4]]$params
#> $kri0001$workflow[[4]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0001$workflow[[4]]$params$vThreshold
#> NULL
#> 
#> $kri0001$workflow[[4]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $kri0001$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0001.yaml"
#> 
#> $kri0001$name
#> [1] "kri0001"
#> 
#> 
#> $kri0002
#> $kri0002$workflow
#> $kri0002$workflow[[1]]
#> $kri0002$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0002$workflow[[1]]$inputs
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[1]]$output
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[1]]$params
#> $kri0002$workflow[[1]]$params$strDomain
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[1]]$params$strColParam
#> [1] "strTreatmentEmergentCol"
#> 
#> $kri0002$workflow[[1]]$params$strValParam
#> [1] "strTreatmentEmergentVal"
#> 
#> 
#> 
#> $kri0002$workflow[[2]]
#> $kri0002$workflow[[2]]$name
#> [1] "FilterDomain"
#> 
#> $kri0002$workflow[[2]]$inputs
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[2]]$output
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[2]]$params
#> $kri0002$workflow[[2]]$params$strDomain
#> [1] "dfAE"
#> 
#> $kri0002$workflow[[2]]$params$strColParam
#> [1] "strSeriousCol"
#> 
#> $kri0002$workflow[[2]]$params$strValParam
#> [1] "strSeriousVal"
#> 
#> 
#> 
#> $kri0002$workflow[[3]]
#> $kri0002$workflow[[3]]$name
#> [1] "AE_Map_Raw"
#> 
#> $kri0002$workflow[[3]]$inputs
#> [1] "dfSUBJ" "dfAE"  
#> 
#> $kri0002$workflow[[3]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0002$workflow[[4]]
#> $kri0002$workflow[[4]]$name
#> [1] "AE_Assess"
#> 
#> $kri0002$workflow[[4]]$inputs
#> [1] "dfInput"
#> 
#> $kri0002$workflow[[4]]$output
#> [1] "lResults"
#> 
#> $kri0002$workflow[[4]]$params
#> $kri0002$workflow[[4]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0002$workflow[[4]]$params$vThreshold
#> NULL
#> 
#> $kri0002$workflow[[4]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $kri0002$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0002.yaml"
#> 
#> $kri0002$name
#> [1] "kri0002"
#> 
#> 
#> $kri0003
#> $kri0003$workflow
#> $kri0003$workflow[[1]]
#> $kri0003$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0003$workflow[[1]]$inputs
#> [1] "dfPD"
#> 
#> $kri0003$workflow[[1]]$output
#> [1] "dfPD"
#> 
#> $kri0003$workflow[[1]]$params
#> $kri0003$workflow[[1]]$params$strDomain
#> [1] "dfPD"
#> 
#> $kri0003$workflow[[1]]$params$strColParam
#> [1] "strImportantCol"
#> 
#> $kri0003$workflow[[1]]$params$strValParam
#> [1] "strNotImportantVal"
#> 
#> 
#> 
#> $kri0003$workflow[[2]]
#> $kri0003$workflow[[2]]$name
#> [1] "PD_Map_Raw"
#> 
#> $kri0003$workflow[[2]]$inputs
#> [1] "dfPD"   "dfSUBJ"
#> 
#> $kri0003$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0003$workflow[[3]]
#> $kri0003$workflow[[3]]$name
#> [1] "PD_Assess"
#> 
#> $kri0003$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $kri0003$workflow[[3]]$output
#> [1] "lResults"
#> 
#> $kri0003$workflow[[3]]$params
#> $kri0003$workflow[[3]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0003$workflow[[3]]$params$vThreshold
#> NULL
#> 
#> $kri0003$workflow[[3]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $kri0003$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0003.yaml"
#> 
#> $kri0003$name
#> [1] "kri0003"
#> 
#> 
#> $kri0004
#> $kri0004$workflow
#> $kri0004$workflow[[1]]
#> $kri0004$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0004$workflow[[1]]$inputs
#> [1] "dfPD"
#> 
#> $kri0004$workflow[[1]]$output
#> [1] "dfPD"
#> 
#> $kri0004$workflow[[1]]$params
#> $kri0004$workflow[[1]]$params$strDomain
#> [1] "dfPD"
#> 
#> $kri0004$workflow[[1]]$params$strColParam
#> [1] "strImportantCol"
#> 
#> $kri0004$workflow[[1]]$params$strValParam
#> [1] "strImportantVal"
#> 
#> 
#> 
#> $kri0004$workflow[[2]]
#> $kri0004$workflow[[2]]$name
#> [1] "PD_Map_Raw"
#> 
#> $kri0004$workflow[[2]]$inputs
#> [1] "dfSUBJ" "dfPD"  
#> 
#> $kri0004$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0004$workflow[[3]]
#> $kri0004$workflow[[3]]$name
#> [1] "PD_Assess"
#> 
#> $kri0004$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $kri0004$workflow[[3]]$output
#> [1] "lResults"
#> 
#> $kri0004$workflow[[3]]$params
#> $kri0004$workflow[[3]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0004$workflow[[3]]$params$vThreshold
#> NULL
#> 
#> $kri0004$workflow[[3]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $kri0004$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0004.yaml"
#> 
#> $kri0004$name
#> [1] "kri0004"
#> 
#> 
#> $kri0005
#> $kri0005$workflow
#> $kri0005$workflow[[1]]
#> $kri0005$workflow[[1]]$name
#> [1] "LB_Map_Raw"
#> 
#> $kri0005$workflow[[1]]$inputs
#> [1] "dfLB"   "dfSUBJ"
#> 
#> $kri0005$workflow[[1]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0005$workflow[[2]]
#> $kri0005$workflow[[2]]$name
#> [1] "LB_Assess"
#> 
#> $kri0005$workflow[[2]]$inputs
#> [1] "dfInput"
#> 
#> $kri0005$workflow[[2]]$output
#> [1] "lResults"
#> 
#> $kri0005$workflow[[2]]$params
#> $kri0005$workflow[[2]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0005$workflow[[2]]$params$vThreshold
#> NULL
#> 
#> $kri0005$workflow[[2]]$params$strMethod
#> [1] "fisher"
#> 
#> 
#> 
#> 
#> $kri0005$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0005.yaml"
#> 
#> $kri0005$name
#> [1] "kri0005"
#> 
#> 
#> $kri0006
#> $kri0006$workflow
#> $kri0006$workflow[[1]]
#> $kri0006$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0006$workflow[[1]]$inputs
#> [1] "dfLB"
#> 
#> $kri0006$workflow[[1]]$output
#> [1] "dfLB"
#> 
#> $kri0006$workflow[[1]]$params
#> $kri0006$workflow[[1]]$params$strDomain
#> [1] "dfLB"
#> 
#> $kri0006$workflow[[1]]$params$strColParam
#> [1] "strGradeCol"
#> 
#> $kri0006$workflow[[1]]$params$strValParam
#> [1] "strGradeHighVal"
#> 
#> 
#> 
#> $kri0006$workflow[[2]]
#> $kri0006$workflow[[2]]$name
#> [1] "LB_Map_Raw"
#> 
#> $kri0006$workflow[[2]]$inputs
#> [1] "dfSUBJ" "dfLB"  
#> 
#> $kri0006$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> 
#> $kri0006$workflow[[3]]
#> $kri0006$workflow[[3]]$name
#> [1] "LB_Assess"
#> 
#> $kri0006$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $kri0006$workflow[[3]]$output
#> [1] "lResults"
#> 
#> $kri0006$workflow[[3]]$params
#> $kri0006$workflow[[3]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0006$workflow[[3]]$params$vThreshold
#> NULL
#> 
#> $kri0006$workflow[[3]]$params$strMethod
#> [1] "fisher"
#> 
#> 
#> 
#> 
#> $kri0006$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0006.yaml"
#> 
#> $kri0006$name
#> [1] "kri0006"
#> 
#> 
#> $kri0007
#> $kri0007$workflow
#> $kri0007$workflow[[1]]
#> $kri0007$workflow[[1]]$name
#> [1] "Disp_Map_Raw"
#> 
#> $kri0007$workflow[[1]]$inputs
#> [1] "dfSUBJ"     "dfSTUDCOMP"
#> 
#> $kri0007$workflow[[1]]$output
#> [1] "dfInput"
#> 
#> $kri0007$workflow[[1]]$params
#> $kri0007$workflow[[1]]$params$strContext
#> [1] "Study"
#> 
#> 
#> 
#> $kri0007$workflow[[2]]
#> $kri0007$workflow[[2]]$name
#> [1] "Disp_Assess"
#> 
#> $kri0007$workflow[[2]]$inputs
#> [1] "dfInput"
#> 
#> $kri0007$workflow[[2]]$output
#> [1] "lResults"
#> 
#> $kri0007$workflow[[2]]$params
#> $kri0007$workflow[[2]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0007$workflow[[2]]$params$vThreshold
#> NULL
#> 
#> $kri0007$workflow[[2]]$params$strMethod
#> [1] "fisher"
#> 
#> 
#> 
#> 
#> $kri0007$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0007.yaml"
#> 
#> $kri0007$name
#> [1] "kri0007"
#> 
#> 
#> $kri0008
#> $kri0008$workflow
#> $kri0008$workflow[[1]]
#> $kri0008$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $kri0008$workflow[[1]]$inputs
#> [1] "dfSDRGCOMP"
#> 
#> $kri0008$workflow[[1]]$output
#> [1] "dfSDRGCOMP"
#> 
#> $kri0008$workflow[[1]]$params
#> $kri0008$workflow[[1]]$params$strDomain
#> [1] "dfSDRGCOMP"
#> 
#> $kri0008$workflow[[1]]$params$strColParam
#> [1] "strTreatmentPhaseCol"
#> 
#> $kri0008$workflow[[1]]$params$strValParam
#> [1] "strTreatmentPhaseVal"
#> 
#> 
#> 
#> $kri0008$workflow[[2]]
#> $kri0008$workflow[[2]]$name
#> [1] "Disp_Map_Raw"
#> 
#> $kri0008$workflow[[2]]$inputs
#> [1] "dfSUBJ"     "dfSDRGCOMP"
#> 
#> $kri0008$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> $kri0008$workflow[[2]]$params
#> $kri0008$workflow[[2]]$params$strContext
#> [1] "Treatment"
#> 
#> 
#> 
#> $kri0008$workflow[[3]]
#> $kri0008$workflow[[3]]$name
#> [1] "Disp_Assess"
#> 
#> $kri0008$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $kri0008$workflow[[3]]$output
#> [1] "lResults"
#> 
#> $kri0008$workflow[[3]]$params
#> $kri0008$workflow[[3]]$params$strGroup
#> [1] "Site"
#> 
#> $kri0008$workflow[[3]]$params$vThreshold
#> NULL
#> 
#> $kri0008$workflow[[3]]$params$strMethod
#> [1] "fisher"
#> 
#> 
#> 
#> 
#> $kri0008$path
#> [1] "/home/runner/work/_temp/Library/gsm/workflow/kri0008.yaml"
#> 
#> $kri0008$name
#> [1] "kri0008"
#> 
#>