Make_Snapshot() ingests data from a variety of sources, and runs KRIs and/or QTLs based on the list provided in lAssessments.

For more context about the inputs and outputs of Make_Snapshot(), refer to the GSM Data Pipeline Vignette, specifically Appendix 2 - Data Model Specifications

Make_Snapshot(
  lMeta = list(config_param = gsm::config_param, config_workflow = gsm::config_workflow,
    meta_params = gsm::meta_param, meta_site = clindata::ctms_site, meta_study =
    clindata::ctms_study, meta_workflow = gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfPD =
    clindata::ctms_protdev, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp %>%
    filter(.data$phase == "Blinded Study Drug Completion"), dfDATACHG =
    clindata::edc_data_points, dfDATAENT = clindata::edc_data_pages, dfQUERY =
    clindata::edc_queries, dfENROLL = clindata::rawplus_enroll),
  lMapping = Read_Mapping(),
  lAssessments = NULL,
  strAnalysisDate = NULL,
  bUpdateParams = FALSE,
  bQuiet = TRUE
)

Arguments

lMeta

list a named list of data frames containing metadata, configuration, and workflow parameters for a given study. See the Data Model Vignette - Appendix 2 - Data Model Specifications for detailed specifications.

lData

list a named list of domain-level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lMapping

list Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined mapping for raw+.

lAssessments

list a named list of metadata defining how each assessment should be run. By default, MakeWorkflowList() imports YAML specifications from inst/workflow.

strAnalysisDate

character date that the data was pulled/wrangled/snapshot. Note: date should be provided in format: YYYY-MM-DD.

bUpdateParams

logical if TRUE, configurable parameters found in lMeta$config_param will overwrite the default values in lMeta$meta_params. Default: FALSE.

bQuiet

logical Suppress warning messages? Default: TRUE.

Value

list

lSnapshot, a named list with a data.frame for each component of the gsm data model.

status_study

status_site

status_workflow

status_param

results_summary

results_analysis

results_bounds

meta_workflow

meta_param

Details

Make Snapshot - create and export data model.

Examples

# run with default testing data
if (FALSE) {
snapshot <- Make_Snapshot()
}