Create interactive timeline for study data

Create interactive timeline for study data

Make_Timeline(study, n_breaks = 10, date_format = "%b\n%Y")

Arguments

study

the table containing study data of interest

n_breaks

the number of breaks to include in the x-axis. NOTE pretty breaks are used to assign break locations and may give more or less that the number of breaks specified.

date_format

The format to apply to the x-axis labels

Value

plot interactive timeline plot.

Examples

Make_Timeline(clindata::ctms_study)