Merge Domain data with subject-level data

Merge Domain data with subject-level data

MergeSubjects(
  dfDomain,
  dfSubjects,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  bQuiet = TRUE
)

Arguments

dfDomain

Subject-level domain data containing one record per participant.

dfSubjects

Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)

strIDCol

name of ID Column - default='SubjectID'

vFillZero

Columns from dfDomain to fill with zeros when no matching row is found in for an ID in dfSubject

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one record per strIDCol.

Examples

MergeSubjects(
  dfDomain = clindata::rawplus_consent,
  dfSubjects = clindata::rawplus_subj,
  strIDCol = "SubjectID"
)
#> # A tibble: 1,297 × 21
#>    Subje…¹ Scree…² SiteID StudyID Scree…³ RandDate   RandF…⁴ FirstStudyDate     
#>    <chr>   <chr>   <chr>  <chr>   <chr>   <date>     <chr>   <dttm>             
#>  1 1277    073     X002X  AA-AA-… Y       2007-07-12 Y       2007-07-12 00:00:00
#>  2 1149    016     X002X  AA-AA-… Y       2005-04-11 Y       2005-04-11 00:00:00
#>  3 1028    145     X002X  AA-AA-… Y       2008-05-18 Y       2008-05-18 00:00:00
#>  4 0650    058     X002X  AA-AA-… Y       2009-02-24 Y       2009-02-24 00:00:00
#>  5 0608    020     X002X  AA-AA-… Y       2012-02-27 Y       2012-02-27 00:00:00
#>  6 0579    092     X002X  AA-AA-… Y       2014-05-15 Y       2014-05-15 00:00:00
#>  7 0407    061     X002X  AA-AA-… Y       2011-03-20 Y       2011-03-20 00:00:00
#>  8 0187    067     X002X  AA-AA-… Y       2014-05-09 Y       2014-05-09 00:00:00
#>  9 0005    093     X002X  AA-AA-… Y       2012-02-19 Y       2012-02-19 00:00:00
#> 10 0535    080     X007X  AA-AA-… Y       2010-05-23 Y       2010-05-24 00:00:00
#> # … with 1,287 more rows, 13 more variables: LastStudyDate <date>,
#> #   TimeOnStudy <dbl>, FirstDoseDate <dttm>, LastDoseDate <date>,
#> #   TimeOnTreatment <dbl>, StudCompletion <chr>, StudDCReason <chr>,
#> #   TrtCompletion <chr>, TrtDCReason <chr>, RegionID <chr>,
#> #   CONSENT_DATE <date>, CONSENT_TYPE <chr>, CONSENT_VALUE <chr>, and
#> #   abbreviated variable names ¹​SubjectID, ²​ScreenID, ³​ScreenFlag, ⁴​RandFlag
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names