Protocol Deviation Assessment

Flag sites that may be over- or under-reporting protocol deviations (PDs).

PD_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "poisson",
  lTags = list(Assessment = "PD"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)

Arguments

dfInput

data.frame Input data, a data frame with one record per subject.

vThreshold

numeric Threshold specification, a vector of length 2 that defaults to c(-5, 5) for strMethod = "poisson" and c(.0001, NA) for strMethod = "wilcoxon".

strMethod

character Statistical model. Valid values include "poisson" (default) and "wilcoxon".

lTags

list Assessment tags, a named list of tags describing the assessment that defaults to list(Assessment="PD"). lTags is returned as part of the assessment (lAssess$lTags) and each tag is added as a column in lAssess$dfSummary.

bChart

logical Generate data visualization? Default: TRUE

bReturnChecks

logical Return input checks from is_mapping_valid? Default: FALSE

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list Assessment, a named list with:

each data frame in the data pipeline

dfInput

dfTransformed, returned by Transform_EventCount()

dfAnalyzed, returned by Analyze_Poisson() or Analyze_Wilcoxon()

dfFlagged, returned by Flag()

dfSummary, returned by Summarize()

assessment metadata

strFunctionName

lParams

lTags

output(s)

chart

Details

The Protocol Deviation Assessment uses the standard GSM data pipeline to flag possible outliers. Additional details regarding the data pipeline and statistical methods are described below.

Data specification

domain	col_key	col_value	vRequired	vUniqueCols

dfInput	strIDCol	SubjectID	TRUE	TRUE

dfInput	strSiteCol	SiteID	TRUE	FALSE

dfInput	strCountCol	Count	TRUE	FALSE

dfInput	strExposureCol	Exposure	TRUE	FALSE

dfInput	strRateCol	Rate	TRUE	FALSE

Statistical assumptions

A Poisson or Wilcoxon model is used to generate estimates and p-values for each site (as specified with the strMethod parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in vThreshold. In the Poisson model, sites with an estimate less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default. For Wilcoxon, sites with p-values less than 0.0001 are flagged by default.

See Analyze_Poisson() and Analyze_Wilcoxon() for additional details about the statistical methods and their assumptions.

Examples

dfInput <- PD_Map_Raw()
#> ℹ Intializing merge of domain and subject data
pd_assessment_poisson <- PD_Assess(dfInput)
pd_assessment_wilcoxon <- PD_Assess(dfInput, strMethod = "wilcoxon")