Evaluates Protocol Deviation (PD) rates to identify sites that may be over- or under-reporting PDs.

PD_Assess_Binary(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "PD_Assess_Binary.yaml", package =
    "gsm")),
  strGroup = "Site",
  nMinDenominator = NULL,
  nConfLevel = NULL,
  bQuiet = TRUE
)

Arguments

dfInput

data.frame Input data, a data frame with one record per subject.

vThreshold

numeric Threshold specification, a numeric vector that defaults to c(-3, -2, 2, 3) for a Normal Approximation (strMethod = "NormalApprox"), c(-7, -5, 5, 7) for a Poisson model (strMethod = "Poisson"), c(0.000895, 0.003059) for a nominal assessment (strMethod = "Identity"), and c(0.01) for a QTL (strMethod = "QTL").

strMethod

character Statistical method. Valid values:

"NormalApprox" (default)

"Poisson"

"Identity"

"QTL"

lMapping

list Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined Protocol Deviation Assessment mapping.

strGroup

character Grouping variable. "Site" (the default) uses the column named in mapping$strSiteCol. Other valid options using the default mapping are "Study", "Country", and "CustomGroup".

nMinDenominator

numeric Specifies the minimum denominator required to return a score and calculate a flag. Default: NULL

nConfLevel

numeric Confidence level for QTL analysis.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list

lData, a named list with:

each data frame in the data pipeline

dfTransformed, returned by Transform_Rate()

dfAnalyzed, returned by Analyze_NormalApprox(), Analyze_Poisson() or Analyze_Identity()

dfFlagged, returned by Flag_NormalApprox(), Flag_Poisson(), or Flag()

dfSummary, returned by Summarize()

dfBounds, returned by Analyze_NormalApprox_PredictBounds() when strMethod == "NormalApprox" or Analyze_Poisson_PredictBounds() when strMethod == "Poisson"

list lCharts, a named list with:

scatter, a ggplot2 object returned by Visualize_Scatter() only when strMethod != "Identity"

barMetric, a ggplot2 object returned by Visualize_Score()

barScore, a ggplot2 object returned by Visualize_Score()

list lChecks, a named list with:

dfInput, a named list returned by is_mapping_valid()

status, a boolean returned by is_mapping_valid()

mapping, a named list that is provided as an argument to the lMapping parameter in PD_Assess_Rate()

spec, a named list used to define variable specifications

Details

Protocol Deviation Assessment (Binary)

The Protocol Deviation Assessment uses the standard GSM data pipeline to flag possible outliers. Additional details regarding the data pipeline and statistical methods are described below.

Data specification

Domain	Column Key	Default Value	Required?	Require Unique Values?	Accept NA/Empty Values?

dfInput	strIDCol	SubjectID	TRUE	TRUE	FALSE

dfInput	strCountCol	Count	TRUE	FALSE	FALSE

dfInput	strGroupCol	SiteID	TRUE	FALSE	FALSE

dfInput	strExposureCol		FALSE	FALSE	TRUE

dfInput	strRateCol		FALSE	FALSE	TRUE

Statistical assumptions

The default function Analyze_NormalApprox applies funnel plots using asymptotic limits based on normal approximation of binomial distribution for the binary outcome, or normal approximation of Poisson distribution for the rate outcome with volume (the sample sizes or total exposure of the sites) to assess data quality and safety.

Alternatively, a Poisson model is used to generate estimates and p-values for each site (as specified with the strMethod parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in vThreshold. Sites with an estimate less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default.

See Analyze_NormalApprox() or Analyze_Poisson() for additional details about the statistical methods and their assumptions.

Examples

dfInput <- PD_Map_Raw_Binary()
pd_assessment_NormalApprox <- PD_Assess_Binary(dfInput)
pd_assessment_poisson <- PD_Assess_Binary(dfInput, strMethod = "Poisson")
pd_assessment_identity <- PD_Assess_Binary(dfInput, strMethod = "Identity")