Query Rate Assessment

Evaluates query rates to identify sites that may be over- or under-reporting queries.

QueryRate_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "QueryRate_Assess.yaml", package =
    "gsm")),
  lLabels = list(workflowid = "", group = strGroup, abbreviation = "QRY", metric =
    "Query Rate", numerator = "Total Queries", denominator = "Total Data Points", model =
    "Normal Approximation", score = "Adjusted Z-Score"),
  strGroup = "Site",
  nMinDenominator = NULL,
  bMakeCharts = FALSE,
  bQuiet = TRUE
)

Arguments

dfInput

data.frame Input data, a data frame with one record per data point.

vThreshold

numeric Threshold specification, a vector of length 4 that defaults to c(-3, -2, 2, 3) for a Normal Approximation, c(-7, -5, 5, 7) for a Poisson model (strMethod = "Poisson") and a vector of length 2 that defaults to c(0.00006, 0.01) for a nominal assessment (strMethod = "Identity").

strMethod

character Statistical method. Valid values:

"NormalApprox" (default)

"Poisson"

"Identity"

lMapping

Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined Adverse Event Assessment mapping.

lLabels

list Labels used to populate chart labels.

strGroup

character Grouping variable. "Site" (the default) uses the column named in mapping$strSiteCol. Other valid options using the default mapping are "Study", "Country", and "CustomGroup".

nMinDenominator

numeric Specifies the minimum denominator required to return a score and calculate a flag. Default: NULL

bMakeCharts

logical Boolean value indicating whether to create charts.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list

lData, a named list with:

each data frame in the data pipeline

dfTransformed, returned by Transform_Rate()

dfAnalyzed, returned by Analyze_NormalApprox(), Analyze_Poisson(), or Analyze_Identity()

dfFlagged, returned by Flag_NormalApprox(), Flag_Poisson(), or Flag()

dfSummary, returned by Summarize()

dfBounds, returned by Analyze_NormalApprox_PredictBounds() or Analyze_Poisson_PredictBounds() when strMethod == "NormalApprox" or strMethod == "Poisson". dfBounds is not returned when using strMethod == "Identity".

list lCharts, a named list with:

scatter, a ggplot2 object returned by Visualize_Scatter() only when strMethod != "Identity"

barMetric, a ggplot2 object returned by Visualize_Score()

barScore, a ggplot2 object returned by Visualize_Score()

list lChecks, a named list with:

dfInput, a named list returned by is_mapping_valid()

status, a boolean returned by is_mapping_valid()

mapping, a named list that is provided as an argument to the lMapping parameter in QueryRate_Assess()

spec, a named list used to define variable specifications

Details

The Query Rate Assessment uses the standard GSM data pipeline to flag possible outliers. Additional details regarding the data pipeline and statistical methods are described below.

Data specification

Domain	Column Key	Default Value	Required?

dfInput	strIDCol	SubjectID	TRUE

dfInput	strCountCol	Count	TRUE

dfInput	strGroupCol	SiteID	TRUE

Statistical assumptions

The default function Analyze_NormalApprox applies funnel plots using asymptotic limits based on normal approximation of binomial distribution for the binary outcome, or normal approximation of Poisson distribution for the rate outcome with volume (the sample sizes or total exposure of the sites) to assess data quality and safety.

Alternatively, a Poisson model is used to generate estimates and p-values for each site (as specified with the strMethod parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in vThreshold. Sites with an estimate less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default.

See Analyze_NormalApprox() or Analyze_Poisson() for additional details about the statistical methods and their assumptions.

Examples

dfInput <- QueryRate_Map_Raw() %>% na.omit()

# Run using normal approximation method (default)
QueryRate_assessment_NormalApprox <- QueryRate_Assess(dfInput)

# Run using Poisson method
QueryRate_assessment_poisson <- QueryRate_Assess(dfInput, strMethod = "Poisson")

# Run using Identity method
QueryRate_assessment_identity <- QueryRate_Assess(dfInput, strMethod = "Identity")