Run a single assessment

Attempts to run a single assessment (lAssessment) using shared data (lData) and metadata (lMapping). Calls RunStep for each item in lAssessment$workflow and saves the results to lAssessment

RunAssessment(lAssessment, lData, lMapping, lTags = NULL, bQuiet = FALSE)

Arguments

lAssessment

list A named list of metadata defining how each assessment should be run. Properties should include: label, tags and workflow

lData

list A named list of domain-level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lMapping

list A named list identifying the columns needed in each data domain.

lTags

list A named list of tags describing the assessment. lTags is returned as part of the assessment (lAssess$lTags) and each tag is added as columns in lassess$dfSummary.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list containing lAssessment with tags, workflow, path, name, lData, lChecks, bStatus, checks, and lResults added based on the results of the execution of assessment$workflow.

Examples

lAssessments <- MakeAssessmentList()
lData <- list(
  dfSUBJ = clindata::rawplus_subj,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_pd,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)
lTags <- list(
  Study = "myStudy"
)
lMapping <- clindata::mapping_rawplus

ae_assessment <- RunAssessment(lAssessments$ae, lData = lData, lMapping = lMapping, lTags = lTags)
#> 
#> ── Initializing `ae` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 3: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 3: `AE_Map_Raw` ──
#> 
#> Preparing parameters for `AE_Map_Raw()` ...
#> Calling `AE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `AE_Map_Raw()` ──
#> 
#> ✔ No issues found for `AE_Map_Raw()`
#> 
#> ── Initializing `AE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 472 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `AE_Map_Raw()` returned output with 1297 rows.
#> ✔ `AE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 3: `AE_Assess` ──
#> 
#> Preparing parameters for `AE_Assess()` ...
#> Calling `AE_Assess()` ...
#> 
#> ── Checking Input Data for `AE_Assess()` ──
#> 
#> ✔ No issues found for `AE_Assess()`
#> 
#> ── Initializing `AE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `AE_Assess()` Successful
#> Saving lResults to `lAssessment`