Run a single step in a workflow

Runs a single step of an assessment workflow. Currently supports Filter, Map, and Assess functions.

RunStep(lStep, lMapping, lData, bQuiet)

Arguments

lStep

list single workflow step (typically defined in lWorkflow$workflow). Should include the name of the function to run (lStep$name), data inputs (lStep$inputs), name of output (lStep$output) and configurable parameters (lStep$params) (if any)

lMapping

list List containing expected columns in each data set.

lData

list a named list of domain level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

bQuiet

logical Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.

Value

list containing the results of the lStep$name function call should contain .$checks

parameter with results from is_mapping_vald for each domain in lStep$inputs.

Examples

lStep <- MakeWorkflowList()[["kri0001"]][["steps"]][[1]]

lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)

ae_step <- RunStep(lStep = lStep, lMapping = lMapping, lData = lData, bQuiet = FALSE)
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on `ae_te %in% c("Y")`.
#> ✔ Filtered on `ae_te %in% c("Y")` to drop 111 rows from 4854 to 4743 rows.