Run a workflow via it's YAML specification.

Attempts to run a single assessment (lWorkflow) using shared data (lData) and metadata (lMapping). Calls RunStep for each item in lWorkflow$workflow and saves the results to lWorkflow.

RunWorkflow(lWorkflow, lData, bReturnData = TRUE, bKeepInputData = FALSE)

Arguments

lWorkflow

list A named list of metadata defining how the workflow should be run.

lData

list A named list of domain-level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

bReturnData

boolean should function return only lData or should meta and steps be included? Default is TRUE.

bKeepInputData

boolean should the input data be returned? Default is FALSE.

Value

list containing objects named: steps, path, name, lData, lChecks, bStatus, lWorkflowChecks, and lResults.

list contains just lData if bReturnData is TRUE, otherwise returns the full lWorkflow object.

Details

Examples

if (FALSE) { # \dontrun{
lAssessments <- MakeWorkflowList("kri0001")
lData <- list(
  dfAE = clindata::rawplus_ae,
  dfCONSENT = clindata::rawplus_consent,
  dfDISP = clindata::rawplus_dm,
  dfIE = clindata::rawplus_ie,
  dfLB = clindata::rawplus_lb,
  dfPD = clindata::ctms_protdev,
  dfSUBJ = clindata::rawplus_dm
)
wf_mapping <- MakeWorkflowList("mapping")
lMapped <- RunWorkflow(wf_mapping, LData)$lData

output <- map(lAssessments, ~RunWorkflow(., lMapped))
} # }