Save QTL analysis results to a directory location. The strPath argument specifies the filepath, including filename, of the current QTL analysis. Running SaveQTL will save a new file with an added row for the single specified QTL analysis. The new file will be named the same as the original file, but will incorporate the run date for version control.

SaveQTL(
  lSnapshot,
  strPath = "./inst/qtl_dummy_data/dummyqtldata.csv",
  bQuiet = TRUE
)

Arguments

lSnapshot

list List returned by RunQTL()

strPath

character Path to historical QTL data.

bQuiet

logical Suppress warning messages? Default: TRUE

Details

SaveQTL

Examples

if (FALSE) {
dispQTL <- RunQTL("qtl0006")
SaveQTL(dispQTL)
}