Convert raw screening data to formatted input data to Screening_Assess().

Screening_Map_Raw(
  dfs = list(dfENROLL = clindata::rawplus_enroll),
  lMapping = gsm::Read_Mapping("rawplus"),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)

Arguments

dfs

list Input data frames:

dfENROLL: data.frame Subject-level study enrollment data with one record per subject. Default: clindata::rawplus_enroll

lMapping

list Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined mapping for raw+.

bReturnChecks

logical Return input checks from is_mapping_valid()? Default: FALSE

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one record per subject, the input to Screening_Assess(). If bReturnChecks is TRUE

Screening_Map_Raw returns a named list with:

df: the data frame described above

lChecks: a named list of check results

Details

Screening Assessment - Raw Mapping

Screening_Map_Raw creates an input dataset for the Screening Assessment Screening_Assess() by identifying Screen Failures (derived from dfENROLL).

Data specification

Domain	Column Key	Default Value	Required?	Accept NA/Empty Values?

dfENROLL	strSiteCol	siteid	TRUE	FALSE

dfENROLL	strIDCol	subjectid	TRUE	FALSE

dfENROLL	strScreenFailCol	enrollyn	TRUE	TRUE

dfENROLL	strScreenFailReasonCol	sfreas	TRUE	TRUE

Examples

# Run with defaults.
dfInput <- Screening_Map_Raw()

# Run with error checking and message log.
dfInput <- Screening_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)
#> 
#> ── Checking Input Data for `Screening_Map_Raw()` ──
#> 
#> ✔ No issues found for `Screening_Map_Raw()`
#> 
#> ── Initializing `Screening_Map_Raw()` ──
#> 
#> ✔ `Screening_Map_Raw()` returned output with 2387 rows.