Run Multiple Assessments on a Study

Attempts to run one or more assessments (lAssessments) using shared data (lData) and metadata (lMapping). By default, the sample rawplus data from the clindata package is used, and all assessments defined in inst/workflow are evaluated. Individual assessments are run using gsm::RunAssessment()

Study_Assess(
  lData = NULL,
  lMapping = NULL,
  lAssessments = NULL,
  lSubjFilters = NULL,
  lTags = list(Study = "myStudy"),
  bQuiet = FALSE
)

Arguments

lData

a named list of domain level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lMapping

a named list identifying the columns needed in each data domain.

lAssessments

a named list of metadata defining how each assessment should be run. By default, MakeAssessmentList() imports YAML specifications from inst/workflow.

lSubjFilters

a named list of parameters to filter subject-level data on.

lTags

a named list of Tags to be passed to each assessment. Default is list(Study="myStudy") could be expanded to include other important metadata such as analysis population or study phase.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list of assessments containing status information and results.

Examples

results <- Study_Assess() # run using defaults
#> 
#> ── Initializing `ae` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 3: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 3: `AE_Map_Raw` ──
#> 
#> Preparing parameters for `AE_Map_Raw()` ...
#> Calling `AE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `AE_Map_Raw()` ──
#> 
#> ✔ No issues found for `AE_Map_Raw()`
#> 
#> ── Initializing `AE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 472 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `AE_Map_Raw()` returned output with 1297 rows.
#> ✔ `AE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 3: `AE_Assess` ──
#> 
#> Preparing parameters for `AE_Assess()` ...
#> Calling `AE_Assess()` ...
#> 
#> ── Checking Input Data for `AE_Assess()` ──
#> 
#> ✔ No issues found for `AE_Assess()`
#> 
#> ── Initializing `AE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `AE_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `consent` assessment ───────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `Consent_Map_Raw` ──
#> 
#> Preparing parameters for `Consent_Map_Raw()` ...
#> Calling `Consent_Map_Raw()` ...
#> 
#> ── Checking Input Data for `Consent_Map_Raw()` ──
#> 
#> ✔ No issues found for `Consent_Map_Raw()`
#> 
#> ── Initializing `Consent_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ✔ `Consent_Map_Raw()` returned output with 1297 rows.
#> ✔ `Consent_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `Consent_Assess` ──
#> 
#> Preparing parameters for `Consent_Assess()` ...
#> Calling `Consent_Assess()` ...
#> 
#> ── Checking Input Data for `Consent_Assess()` ──
#> 
#> ✔ No issues found for `Consent_Assess()`
#> 
#> ── Initializing `Consent_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> `Score` column created from `KRI`.
#> `ScoreLabel` column created from `KRILabel`.
#> ℹ No analysis function used. `dfTransformed` copied directly to `dfAnalyzed` with added `ScoreLabel` column.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Count()` created a chart.
#> ✔ `Consent_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `ie` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `IE_Map_Raw` ──
#> 
#> Preparing parameters for `IE_Map_Raw()` ...
#> Calling `IE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `IE_Map_Raw()` ──
#> 
#> ✔ No issues found for `IE_Map_Raw()`
#> 
#> ── Initializing `IE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 13 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `IE_Map_Raw()` returned output with 1297 rows.
#> ✔ `IE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `IE_Assess` ──
#> 
#> Preparing parameters for `IE_Assess()` ...
#> Calling `IE_Assess()` ...
#> 
#> ── Checking Input Data for `IE_Assess()` ──
#> 
#> ✔ No issues found for `IE_Assess()`
#> 
#> ── Initializing `IE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> `Score` column created from `KRI`.
#> `ScoreLabel` column created from `KRILabel`.
#> ℹ No analysis function used. `dfTransformed` copied directly to `dfAnalyzed` with added `ScoreLabel` column.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Count()` created a chart.
#> ✔ `IE_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `importantpd` assessment ───────────────────────────────────────
#> 
#> ── Workflow Step 1 of 3: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfPD domain
#> Filtering on PD_IMPORTANT_FLAG == Y
#> ✔ Filtered on `PD_IMPORTANT_FLAG=Y`, to drop 3976 rows from 4459 to 483 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfPD to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 3: `PD_Map_Raw` ──
#> 
#> Preparing parameters for `PD_Map_Raw()` ...
#> Calling `PD_Map_Raw()` ...
#> 
#> ── Checking Input Data for `PD_Map_Raw()` ──
#> 
#> ✔ No issues found for `PD_Map_Raw()`
#> 
#> ── Initializing `PD_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 953 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `PD_Map_Raw()` returned output with 1297 rows.
#> ✔ `PD_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 3: `PD_Assess` ──
#> 
#> Preparing parameters for `PD_Assess()` ...
#> Calling `PD_Assess()` ...
#> 
#> ── Checking Input Data for `PD_Assess()` ──
#> 
#> ✔ No issues found for `PD_Assess()`
#> 
#> ── Initializing `PD_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `PD_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `pd` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `PD_Map_Raw` ──
#> 
#> Preparing parameters for `PD_Map_Raw()` ...
#> Calling `PD_Map_Raw()` ...
#> 
#> ── Checking Input Data for `PD_Map_Raw()` ──
#> 
#> ✔ No issues found for `PD_Map_Raw()`
#> 
#> ── Initializing `PD_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 216 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `PD_Map_Raw()` returned output with 1297 rows.
#> ✔ `PD_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `PD_Assess` ──
#> 
#> Preparing parameters for `PD_Assess()` ...
#> Calling `PD_Assess()` ...
#> 
#> ── Checking Input Data for `PD_Assess()` ──
#> 
#> ✔ No issues found for `PD_Assess()`
#> 
#> ── Initializing `PD_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `PD_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `sae` assessment ───────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 4: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 4: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_SERIOUS == Yes
#> ✔ Filtered on `AE_SERIOUS=Yes`, to drop 2385 rows from 2456 to 71 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 4: `AE_Map_Raw` ──
#> 
#> Preparing parameters for `AE_Map_Raw()` ...
#> Calling `AE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `AE_Map_Raw()` ──
#> 
#> ✔ No issues found for `AE_Map_Raw()`
#> 
#> ── Initializing `AE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 1238 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `AE_Map_Raw()` returned output with 1297 rows.
#> ✔ `AE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 4 of 4: `AE_Assess` ──
#> 
#> Preparing parameters for `AE_Assess()` ...
#> Calling `AE_Assess()` ...
#> 
#> ── Checking Input Data for `AE_Assess()` ──
#> 
#> ✔ No issues found for `AE_Assess()`
#> 
#> ── Initializing `AE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `AE_Assess()` Successful
#> Saving lResults to `lAssessment`