Make Assessment overview table

Make overview table with one row per assessment and one column per site showing flagged assessments.

Study_Table(
  dfFindings,
  bFormat = TRUE,
  bShowCounts = TRUE,
  bShowSiteScore = TRUE,
  vSiteScoreThreshold = 1,
  bColCollapse = TRUE
)

Arguments

dfFindings

dataframe containing one or more stacked findings. Findings are one record per assessment per site and have the following columns: Assessment, Label, SiteID, N, PValue, Flag. PValue is ignored in the summary table.

bFormat

Use html-friendly icons in table cells. -1 is converted to a down arrow. 1 is converted to an up arrow. 0 is not shown. Other values are left as is.

bShowCounts

Show site counts? Uses first value of N for each site given in dfFindings.

bShowSiteScore

Show a "Score" row with total number of flagged assessments for each site. TODO: add method for custom scoring in future release)

vSiteScoreThreshold

Hide sites with a site score less than this value (1 by default).

bColCollapse

Combine the Assessment and Label columns into a single "Title Column"

Value

data.frame Returns a data.frame giving assessment status (rows) by Site (column)

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(purrr)
results <- Study_Assess() %>%
  purrr::map(~ .x$lResults) %>%
  compact() %>%
  purrr::map_df(~ .x$dfSummary)
#> 
#> ── Initializing `ae` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 3: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 3: `AE_Map_Raw` ──
#> 
#> Preparing parameters for `AE_Map_Raw()` ...
#> Calling `AE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `AE_Map_Raw()` ──
#> 
#> ✔ No issues found for `AE_Map_Raw()`
#> 
#> ── Initializing `AE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 472 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `AE_Map_Raw()` returned output with 1297 rows.
#> ✔ `AE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 3: `AE_Assess` ──
#> 
#> Preparing parameters for `AE_Assess()` ...
#> Calling `AE_Assess()` ...
#> 
#> ── Checking Input Data for `AE_Assess()` ──
#> 
#> ✔ No issues found for `AE_Assess()`
#> 
#> ── Initializing `AE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `AE_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `consent` assessment ───────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `Consent_Map_Raw` ──
#> 
#> Preparing parameters for `Consent_Map_Raw()` ...
#> Calling `Consent_Map_Raw()` ...
#> 
#> ── Checking Input Data for `Consent_Map_Raw()` ──
#> 
#> ✔ No issues found for `Consent_Map_Raw()`
#> 
#> ── Initializing `Consent_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ✔ `Consent_Map_Raw()` returned output with 1297 rows.
#> ✔ `Consent_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `Consent_Assess` ──
#> 
#> Preparing parameters for `Consent_Assess()` ...
#> Calling `Consent_Assess()` ...
#> 
#> ── Checking Input Data for `Consent_Assess()` ──
#> 
#> ✔ No issues found for `Consent_Assess()`
#> 
#> ── Initializing `Consent_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> `Score` column created from `KRI`.
#> `ScoreLabel` column created from `KRILabel`.
#> ℹ No analysis function used. `dfTransformed` copied directly to `dfAnalyzed` with added `ScoreLabel` column.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Count()` created a chart.
#> ✔ `Consent_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `ie` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `IE_Map_Raw` ──
#> 
#> Preparing parameters for `IE_Map_Raw()` ...
#> Calling `IE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `IE_Map_Raw()` ──
#> 
#> ✔ No issues found for `IE_Map_Raw()`
#> 
#> ── Initializing `IE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 13 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `IE_Map_Raw()` returned output with 1297 rows.
#> ✔ `IE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `IE_Assess` ──
#> 
#> Preparing parameters for `IE_Assess()` ...
#> Calling `IE_Assess()` ...
#> 
#> ── Checking Input Data for `IE_Assess()` ──
#> 
#> ✔ No issues found for `IE_Assess()`
#> 
#> ── Initializing `IE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> `Score` column created from `KRI`.
#> `ScoreLabel` column created from `KRILabel`.
#> ℹ No analysis function used. `dfTransformed` copied directly to `dfAnalyzed` with added `ScoreLabel` column.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Count()` created a chart.
#> ✔ `IE_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `importantpd` assessment ───────────────────────────────────────
#> 
#> ── Workflow Step 1 of 3: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfPD domain
#> Filtering on PD_IMPORTANT_FLAG == Y
#> ✔ Filtered on `PD_IMPORTANT_FLAG=Y`, to drop 3976 rows from 4459 to 483 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfPD to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 3: `PD_Map_Raw` ──
#> 
#> Preparing parameters for `PD_Map_Raw()` ...
#> Calling `PD_Map_Raw()` ...
#> 
#> ── Checking Input Data for `PD_Map_Raw()` ──
#> 
#> ✔ No issues found for `PD_Map_Raw()`
#> 
#> ── Initializing `PD_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 953 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `PD_Map_Raw()` returned output with 1297 rows.
#> ✔ `PD_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 3: `PD_Assess` ──
#> 
#> Preparing parameters for `PD_Assess()` ...
#> Calling `PD_Assess()` ...
#> 
#> ── Checking Input Data for `PD_Assess()` ──
#> 
#> ✔ No issues found for `PD_Assess()`
#> 
#> ── Initializing `PD_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `PD_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `pd` assessment ────────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 2: `PD_Map_Raw` ──
#> 
#> Preparing parameters for `PD_Map_Raw()` ...
#> Calling `PD_Map_Raw()` ...
#> 
#> ── Checking Input Data for `PD_Map_Raw()` ──
#> 
#> ✔ No issues found for `PD_Map_Raw()`
#> 
#> ── Initializing `PD_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 216 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `PD_Map_Raw()` returned output with 1297 rows.
#> ✔ `PD_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 2: `PD_Assess` ──
#> 
#> Preparing parameters for `PD_Assess()` ...
#> Calling `PD_Assess()` ...
#> 
#> ── Checking Input Data for `PD_Assess()` ──
#> 
#> ✔ No issues found for `PD_Assess()`
#> 
#> ── Initializing `PD_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `PD_Assess()` Successful
#> Saving lResults to `lAssessment`
#> 
#> ── Initializing `sae` assessment ───────────────────────────────────────────────
#> 
#> ── Workflow Step 1 of 4: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 2 of 4: `FilterDomain` ──
#> 
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_SERIOUS == Yes
#> ✔ Filtered on `AE_SERIOUS=Yes`, to drop 2385 rows from 2456 to 71 rows.
#> ✔ `FilterDomain()` Successful
#> Saving dfAE to `lAssessment$lData`
#> 
#> ── Workflow Step 3 of 4: `AE_Map_Raw` ──
#> 
#> Preparing parameters for `AE_Map_Raw()` ...
#> Calling `AE_Map_Raw()` ...
#> 
#> ── Checking Input Data for `AE_Map_Raw()` ──
#> 
#> ✔ No issues found for `AE_Map_Raw()`
#> 
#> ── Initializing `AE_Map_Raw()` ──
#> 
#> ℹ Intializing merge of domain and subject data
#> ℹ 1238 ID(s) in subject data not found in domain data.
#> These participants will have 0s imputed for the following domain data columns: Count.
#> NA's will be imputed for all other columns.
#> ✔ `AE_Map_Raw()` returned output with 1297 rows.
#> ✔ `AE_Map_Raw()` Successful
#> Saving dfInput to `lAssessment$lData`
#> 
#> ── Workflow Step 4 of 4: `AE_Assess` ──
#> 
#> Preparing parameters for `AE_Assess()` ...
#> Calling `AE_Assess()` ...
#> 
#> ── Checking Input Data for `AE_Assess()` ──
#> 
#> ✔ No issues found for `AE_Assess()`
#> 
#> ── Initializing `AE_Assess()` ──
#> 
#> Input data has 1297 rows.
#> ✔ `Transform_EventCount()` returned output with 175 rows.
#> ℹ Fitting log-linked Poisson generalized linear model of [ TotalCount ] ~ [ log( TotalExposure ) ].
#> ✔ `Analyze_Poisson()` returned output with 175 rows.
#> ✔ `Flag()` returned output with 175 rows.
#> ✔ `Summarize()` returned output with 175 rows.
#> ✔ `Visualize_Scatter()` created a chart.
#> ✔ `AE_Assess()` Successful
#> Saving lResults to `lAssessment`

Study_Table(results)
#> $df_summary
#> # A tibble: 12 × 43
#>    Title X055X X086X X050X X140X X180X X054X X154X X009X X164X X102X X090X X126X
#>    <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Numb… "43"  "33"  "27"  "25"  "10"  "9"   "7"   "6"   "5"   "69"  "36"  "28" 
#>  2 Score "2"   "2"   "2"   "2"   "2"   "2"   "2"   "2"   "2"   "1"   "1"   "1"  
#>  3 Safe… "<sv… "<sv… ""    "<sv… "<sv… "<sv… "<sv… "<sv… "<sv… "<sv… ""    "<sv…
#>  4 --AEs "<sv… "<sv… " "   "<sv… "<sv… "<sv… "<sv… "<sv… "<sv… "<sv… " "   "<sv…
#>  5 --AE… " "   " "   " "   " "   " "   " "   " "   " "   " "   " "   " "   " "  
#>  6 Cons… "<sv… "<sv… ""    ""    ""    "<sv… ""    ""    "<sv… ""    ""    ""   
#>  7 --Co… "<sv… "<sv… " "   " "   " "   "<sv… " "   " "   "<sv… " "   " "   " "  
#>  8 IE    ""    ""    ""    ""    ""    ""    ""    ""    ""    ""    ""    ""   
#>  9 --IE  " "   " "   " "   " "   " "   " "   " "   " "   " "   " "   " "   " "  
#> 10 PD    ""    ""    "<sv… "<sv… "<sv… ""    "<sv… "<sv… ""    ""    "<sv… ""   
#> 11 --Im… " "   " "   "<sv… " "   " "   " "   " "   " "   " "   " "   " "   " "  
#> 12 --PD  " "   " "   "<sv… "<sv… "<sv… " "   "<sv… "<sv… " "   " "   "<sv… " "  
#> # … with 30 more variables: X192X <chr>, X013X <chr>, X168X <chr>, X236X <chr>,
#> #   X068X <chr>, X033X <chr>, X081X <chr>, X129X <chr>, X018X <chr>,
#> #   X235X <chr>, X037X <chr>, X159X <chr>, X173X <chr>, X204X <chr>,
#> #   X038X <chr>, X100X <chr>, X094X <chr>, X097X <chr>, X143X <chr>,
#> #   X166X <chr>, X174X <chr>, X183X <chr>, X185X <chr>, X224X <chr>,
#> #   X110X <chr>, X117X <chr>, X179X <chr>, X120X <chr>, X132X <chr>,
#> #   X145X <chr>
#> 
#> $footnote
#> [1] "Note: Data not shown for 133 site(s) with site score less than 1"
#>