Make Summary Data Frame

Create a concise summary of assessment results that is easy to aggregate across assessments

Usage

Summarize(dfFlagged, nMinDenominator = NULL)

Arguments

dfFlagged

data.frame in format produced by Flag().

nMinDenominator

numeric Specifies the minimum denominator required to return a score and calculate a flag. Default: NULL

Value

Simplified finding data.frame with columns for GroupID, GroupType, Metric, Score, Flag when associated with a workflow.

Details

Summarize supports the input data (dfFlagged) from the Flag function.

Data Specification

(dfFlagged) has the following required columns:

GroupID - Group ID

GroupLevel - Group Type

Flag - Flagging value of -1, 0, or 1

Score - Column from analysis results.

Examples

dfTransformed <- Transform_Rate(analyticsInput)

dfAnalyzed <- Analyze_Poisson(dfTransformed)
#> ℹ Fitting log-linked Poisson generalized linear model of [ Numerator ] ~ [ log( Denominator ) ].

dfFlagged <- Flag(dfAnalyzed, vThreshold = c(-5, 5))

dfSummary <- Summarize(dfFlagged)