Convert from input data format to needed input format to derive KRI for an Assessment. Calculate site-level count.

Transform_Count(dfInput, strCountCol, strGroupCol = "SiteID")

Arguments

dfInput

A data.frame with one record per subject.

strCountCol

Required. Numerical or logical. Column to be counted.

strGroupCol

character Name of column for grouping variable. Default: "SiteID"

Value

data.frame with one row per site with columns GroupID, TotalCount, and Metric.

Details

Transform Count

This function transforms data to prepare it for the analysis step. It is currently only sourced for the Consent and IE Assessments.

Data Specification

The input data (dfInput) for the Consent and IE Assessments is typically created using any of these functions: Consent_Map_Raw IE_Map_Raw

(dfInput) must include the columns specified by strCountCol and strGroupCol. Required columns include:

SiteID - Site ID

StudyID - Study ID

CustomGroupID - Custom Group ID

Count - Number of events of interest; the actual name of this column is specified by the parameter strCountCol.

The input data has one or more rows per site. Transform_Count() sums strCountCol for a TotalCount for each site. Metric is set to TotalCount to be used downstream in the workflow.

Examples

dfInput <- Consent_Map_Raw()
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")