Transform - Rate

Calculate a site-level Rate

Transform_Rate(
  dfInput,
  strNumeratorCol,
  strDenominatorCol = NULL,
  strGroupCol = "SiteID",
  bQuiet = TRUE
)

Arguments

dfInput

A data.frame with one record per person.

strNumeratorCol

Required. Numerical or logical. Column to be counted.

strDenominatorCol

Optional. Numerical Exposure column.

strGroupCol

character Required. Name of column for grouping variable. Default: "SiteID"

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one row per site with columns SiteID, TotalCount with additional columns Exposure and Rate if strExposureCol is used.

Details

This function transforms data to prepare it for the Analysis step

Data Specification

The input data (dfInput) for the AE Assessment is typically created using any of these functions: AE_Map_Raw AE_Map_Adam Disp_Map_Raw LB_Map_Raw PD_Map_Raw

(dfInput) must include the columns specified by strNumeratorCol, strDenominatorCol and strGroupCol

Examples

dfInput <- AE_Map_Raw()
dfTransformed <- Transform_Rate(dfInput,
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)