Group-level visualization of group-level results

Visualize_Scatter(
  dfResults,
  dfBounds = NULL,
  strGroupCol = NULL,
  strGroupLabel = NULL,
  strUnit = "days",
  vColors = c("#999999", "#FADB14", "#FF4D4F")
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate.

dfBounds

data.frame data.frame giving prediction bounds for range of denominator in dfResults.

strGroupCol

character name of stratification column for facet wrap Default: NULL

strGroupLabel

character name of group, used for labeling axes. Default: NULL

strUnit

character exposure time unit. Default: days

vColors

character vector of hex colors for plotting boundaries/thresholds. Index 1: mean; index 2: first threshold boundary; index 3: second threshold boundary.

Value

group-level plot object.

Examples

if (FALSE) { # \dontrun{
lData <- UseClindata(
  list(
    "dfSUBJ" = "clindata::rawplus_dm",
    "dfAE" = "clindata::rawplus_ae",
    "dfPD" = "clindata::ctms_protdev",
    "dfCONSENT" = "clindata::rawplus_consent",
    "dfIE" = "clindata::rawplus_ie",
    "dfLB" = "clindata::rawplus_lb",
    "dfSTUDCOMP" = "clindata::rawplus_studcomp",
    "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp %>%
          dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
    "dfDATACHG" = "clindata::edc_data_points",
    "dfDATAENT" = "clindata::edc_data_pages",
    "dfQUERY" = "clindata::edc_queries",
    "dfENROLL" = "clindata::rawplus_enroll"
  )
)
wf_mapping <- MakeWorkflowList("mapping")
ae <- MakeWorkflowList(strNames = "kri0001")
lMapped <- RunWorkflow(wf_mapping, lData)$lData
SafetyAE <- map(ae, ~ RunWorkflow(., lMapped))
dfBounds <- Analyze_Poisson_PredictBounds(SafetyAE$kri0001$lData$dfTransformed, c(-5, 5))
Visualize_Scatter(SafetyAE$kri0001$lData$dfSummary, dfBounds)
} # }