A widget that displays a group-level bar chart based on the output of a KRI analysis. Bar charts are provided by default in any Assess function, and are suffixed with "JS" to indicate that they are an htmlwidget ported from the rbm-viz JavaScript library.

Widget_BarChart(
  results = NULL,
  workflow = list(),
  threshold = NULL,
  yaxis = "score",
  selectedGroupIDs = NULL,
  addSiteSelect = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  siteSelectLabelValue = NULL
)

Arguments

results

data with column names:

studyid

workflowid

groupid

numerator

denominator

metric

score

flag

workflow

configuration data with columns:

workflow

gsm_version

group

metric

numerator

denominator

outcome

model

score

data_inputs

data_filters

threshold

a one row data frame containing columns:

workflowid

gsm_version

param

index

default

configurable

yaxis

either 'score' or 'metric'

selectedGroupIDs

group IDs to highlight, NULL by default, can be a single site or a vector.

addSiteSelect

logical add a dropdown to highlight sites? Default: TRUE.

width

width of widget, full screen by default

height

height of widget, calculated based on width

elementId

ID of container HTML element

siteSelectLabelValue

Label used to populate the HTML drop-down menu. Constructed as: 'Highlighted siteSelectLabelValue: '.

Details

KRI Bar Chart

Examples

ae <- AE_Map_Raw()

ae_transform <- Transform_Rate(
  dfInput = ae,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

ae_analyze <- Analyze_NormalApprox(
  dfTransformed = ae_transform,
  strType = "rate"
)

ae_flag <- Flag_NormalApprox(
  ae_analyze,
  vThreshold = c(-3, -2, 2, 3)
)

ae_summary <- Summarize(
  ae_flag
)

dfConfig <- MakeDfConfig(
  strMethod = "NormalApprox",
  strGroup = "Site",
  strAbbreviation = "AE",
  strMetric = "Adverse Event Rate",
  strNumerator = "Adverse Events",
  strDenominator = "Days on Study",
  vThreshold = c(-3, -2, 2, 3)
)

plot <- Widget_BarChart(
  results = ae_summary,
  workflow = dfConfig,
  yaxis = "metric",
  elementId = "aeAssessMetric"
)