Flag Over Time Widget

A widget that generates a table of flags over time using Report_FlagOverTime().

Usage

Widget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country"),
  strFootnote = NULL
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting. Created by passing an lWorkflow object to MakeMetric(). Expected columns: File, MetricID, Group, Abbreviation, Metric, Numerator, Denominator, Model, Score, and Threshold. For more details see the Data Model vignette: vignette("DataModel", package = "gsm").

strGroupLevel

character Value for the group level. Default: "Site".

strFootnote

character Text to insert for figure

Examples

reportingResultsSubset <- dplyr::filter(
  reportingResults,
  GroupID %in% head(unique(reportingResults$GroupID))
)
Widget_FlagOverTime(
  dfResults = reportingResultsSubset,
  dfMetrics = reportingMetrics
)