Group Overview Widget

A widget that generates a group overview table of group-level metric results across one or more metrics.

Widget_GroupOverview(
  dfResults,
  dfMetrics = NULL,
  dfGroups = NULL,
  strGroupLevel = NULL,
  strGroupSubset = "red",
  strGroupLabelKey = "InvestigatorLastName",
  bDebug = FALSE
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting. Created by passing an lWorkflow object to MakeMetric(). Expected columns: File, MetricID, Group, Abbreviation, Metric, Numerator, Denominator, Model, Score, and Threshold. For more details see the Data Pipeline vignette: vignette("DataPipeline", package = "gsm").

dfGroups

data.frame Group metadata.

strGroupLevel

character Value for the group level. Default: NULL and taken from dfMetrics$GroupLevel if available.

strGroupSubset

character Subset of groups to include in the table. Default: 'red'. Options:

'all': All groups.

'red': Groups with 1+ red flags.

'red/amber': Groups with 1+ red/amber flag.

'amber': Groups with 1+ amber flag.

strGroupLabelKey

character Value for the group label key. Default: 'InvestigatorLastName'.

bDebug

logical Print debug messages? Default: FALSE.

Details

Examples

if (FALSE) { # \dontrun{
strGroupLevel <- 'Site'

lWorkflows <- MakeWorkflowList()
lMetricWorkflows <- lWorkflows[ grepl('^kri', names(lWorkflows) ) ]

lDataRaw <- gsm::UseClindata(
    list(
        "dfSUBJ" = "clindata::rawplus_dm",
        "dfAE" = "clindata::rawplus_ae",
        "dfPD" = "clindata::ctms_protdev",
        "dfLB" = "clindata::rawplus_lb",
        "dfSTUDCOMP" = "clindata::rawplus_studcomp",
        "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp %>%
            dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
        "dfDATACHG" = "clindata::edc_data_points",
        "dfDATAENT" = "clindata::edc_data_pages",
        "dfQUERY" = "clindata::edc_queries",
        "dfENROLL" = "clindata::rawplus_enroll"
    )
)

lDataMapped <- RunWorkflow(lWorkflows$data_mapping, lDataRaw)

lResults <- map(
    lMetricWorkflows,
    ~ RunWorkflow(.x, lDataMapped)
)

dfResults <- lResults %>% imap_dfr(~ {
    data <- .x$dfSummary
    data$MetricID <- .y
    data$GroupLevel <- strGroupLevel
    data
})

dfGroups <- bind_rows(
    "SELECT pi_number as GroupID, site_status as Status, pi_first_name as InvestigatorFirstName, pi_last_name as InvestigatorLastName, city as City, state as State, country as Country, * FROM df" %>%
        RunQuery(clindata::ctms_site) %>%
        MakeLongMeta('Site'),
    "SELECT invid as GroupID, COUNT(DISTINCT subjectid) as ParticipantCount, COUNT(DISTINCT invid) as SiteCount FROM df GROUP BY invid" %>%
        RunQuery(lData$dfEnrolled) %>%
        MakeLongMeta('Site'),
    "SELECT country as GroupID, COUNT(DISTINCT subjectid) as ParticipantCount, COUNT(DISTINCT invid) as SiteCount FROM df GROUP BY country" %>%
        RunQuery(lData$dfEnrolled) %>%
        MakeLongMeta('Country')
)

dfMetrics <- lMetricWorkflows %>%
    map_dfr(~ .x$meta)

Widget_GroupOverview(
  dfResults,
  dfMetrics,
  dfGroups
)
} # }