Site Overview Widget

A widget that generates a site overview table of group-level metric results across one or more metrics.

Widget_SiteOverview(
  dfSummary,
  dfMetrics = gsm::meta_workflow,
  dfGroups = example_groups(),
  strGroupSubset = "red",
  bDebug = FALSE
)

Arguments

dfSummary

data.frame Output of Summarize().

dfMetrics

list Metric metadata, captured at the top of metric workflows and returned by MakeMetricInfo().

dfGroups

data.frame Group metadata.

strGroupSubset

character Subset of groups to include in the table. Default: 'red'. Options:

'all': All groups.

'red': Groups with 1+ red flags.

'red/amber': Groups with 1+ red/amber flag.

'amber': Groups with 1+ amber flag.

bDebug

logical Print debug messages? Default: FALSE.

Details

Examples

if (FALSE) { # \dontrun{
data_raw <- UseClindata(
    list(
        "dfSUBJ" = "clindata::rawplus_dm",
        "dfAE" = "clindata::rawplus_ae",
        "dfPD" = "clindata::ctms_protdev",
        "dfLB" = "clindata::rawplus_lb",
        "dfSTUDCOMP" = "clindata::rawplus_studcomp",
        "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp %>%
      dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
        "dfDATACHG" = "clindata::edc_data_points",
        "dfDATAENT" = "clindata::edc_data_pages",
        "dfQUERY" = "clindata::edc_queries",
        "dfENROLL" = "clindata::rawplus_enroll"
    )
)

workflows <- MakeWorkflowList()

data_mapped <- RunWorkflow(workflows$mapping, data_raw)$lData

results <- workflows[ grepl('^kri', names(workflows) ) ] %>%
    map(~ RunWorkflow(.x, data_mapped))

dfSummary <- results %>% map_dfr(~ {
    data <- .x$lData$dfSummary
    data$MetricID <- .x$name
    data
})

counts <- RunWorkflow(workflows$counts, data_mapped)

# TODO: use [ clindata::rawplus_dm$invid ] and [ clindata::ctms_site$pi_number ] instead of
# [ clindata::rawplus_dm$siteid ] and [ clindata::ctms_site$site_num ].
dfGroups <- clindata::ctms_site %>%
  dplyr::left_join(
    clindata::rawplus_dm %>%
      dplyr::group_by(siteid) %>%
      dplyr::tally(name = "enrolled_participants"),
    c('site_num' = 'siteid')
  ) %>%
  dplyr::rename(
    SiteID = site_num,
    status = site_status
  )

dfMetrics <- results %>% map_dfr(~ {
    metric <- .x$meta
    metric$vThreshold <- paste(metric$vThreshold, collapse = ',')
    metric
})

Widget_SiteOverview(
  dfSummary,
  dfMetrics,
  dfGroups
)
} # }