Check that a data frame contains columns and fields specified in mapping

Check that a data frame contains columns and fields specified in mapping

is_mapping_valid(
  df,
  mapping,
  vRequiredParams = NULL,
  bKeepAllParams = FALSE,
  vUniqueCols = NULL,
  vNACols = NULL,
  bQuiet = TRUE
)

Arguments

df

data.frame to compare to mapping object.

mapping

named list specifying expected columns and fields in df.

vRequiredParams

character vector of names that must be present in mapping.

bKeepAllParams

Should params not included in vRequiredParams be evaluated? Default is false.

vUniqueCols

list of columns expected to be unique. default = NULL (none).

vNACols

list of columns where na values are acceptable default = NULL (none).

bQuiet

Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.

Examples

subj_mapping <- list(strIDCol = "SubjectID",
                    strSiteCol = "SiteID",
                    strExposureCol = "TimeOnTreatment")

is_mapping_valid(df = clindata::rawplus_subj,
                 mapping = subj_mapping,
                 vUniqueCols = "SUBJID",
                 vRequiredParams = c("strIDCol", "strSiteCol", "strExposureCol"))
#> $status
#> [1] TRUE
#> 
#> $tests_if
#> $tests_if$is_data_frame
#> $tests_if$is_data_frame$status
#> [1] TRUE
#> 
#> $tests_if$is_data_frame$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_required_params
#> $tests_if$has_required_params$status
#> [1] TRUE
#> 
#> $tests_if$has_required_params$warning
#> [1] NA
#> 
#> 
#> $tests_if$mapping_is_list
#> $tests_if$mapping_is_list$status
#> [1] TRUE
#> 
#> $tests_if$mapping_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mappings_are_character
#> $tests_if$mappings_are_character$status
#> [1] TRUE
#> 
#> $tests_if$mappings_are_character$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_expected_columns
#> $tests_if$has_expected_columns$status
#> [1] TRUE
#> 
#> $tests_if$has_expected_columns$warning
#> [1] NA
#> 
#> 
#> $tests_if$columns_have_na
#> $tests_if$columns_have_na$status
#> [1] TRUE
#> 
#> $tests_if$columns_have_na$warning
#> [1] NA
#> 
#> 
#> $tests_if$columns_have_empty_values
#> $tests_if$columns_have_empty_values$status
#> [1] TRUE
#> 
#> $tests_if$columns_have_empty_values$warning
#> [1] NA
#> 
#> 
#> $tests_if$cols_are_unique
#> $tests_if$cols_are_unique$status
#> [1] TRUE
#> 
#> $tests_if$cols_are_unique$warning
#> [1] NA
#> 
#> 
#> 

subj_mapping$not_a_col <- "nope"

is_mapping_valid(df = clindata::rawplus_subj,
                 mapping = subj_mapping,
                 vUniqueCols = "SUBJID",
                 vRequiredParams = c("strIDCol", "strSiteCol", "strExposureCol"))
#> $status
#> [1] TRUE
#> 
#> $tests_if
#> $tests_if$is_data_frame
#> $tests_if$is_data_frame$status
#> [1] TRUE
#> 
#> $tests_if$is_data_frame$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_required_params
#> $tests_if$has_required_params$status
#> [1] TRUE
#> 
#> $tests_if$has_required_params$warning
#> [1] NA
#> 
#> 
#> $tests_if$mapping_is_list
#> $tests_if$mapping_is_list$status
#> [1] TRUE
#> 
#> $tests_if$mapping_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mappings_are_character
#> $tests_if$mappings_are_character$status
#> [1] TRUE
#> 
#> $tests_if$mappings_are_character$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_expected_columns
#> $tests_if$has_expected_columns$status
#> [1] TRUE
#> 
#> $tests_if$has_expected_columns$warning
#> [1] NA
#> 
#> 
#> $tests_if$columns_have_na
#> $tests_if$columns_have_na$status
#> [1] TRUE
#> 
#> $tests_if$columns_have_na$warning
#> [1] NA
#> 
#> 
#> $tests_if$columns_have_empty_values
#> $tests_if$columns_have_empty_values$status
#> [1] TRUE
#> 
#> $tests_if$columns_have_empty_values$warning
#> [1] NA
#> 
#> 
#> $tests_if$cols_are_unique
#> $tests_if$cols_are_unique$status
#> [1] TRUE
#> 
#> $tests_if$cols_are_unique$warning
#> [1] NA
#> 
#> 
#>