Workflow Metadata

A dataset containing the metadata describing the workflows available in GSM. Each row describes a workflow defined in inst/workflows.

meta_workflow

Format

A data frame with 8 rows and 12 variables:

workflowid

Workflow ID

gsm_version

Version of gsm that the workflow will be run on.

group

Description of grouping variable: Site, Study, or CustomGroup.

metric

Description of the assessment.

numerator

The numerator used to calculate the KRI/QTL.

denominator

The denominator used to calculate the KRI/QTL.

outcome

Description of the outcome. Either Rate or Percent.

model

Statistical model that is used for the assessment.

score

Description of the statistical score that is derived. E.g., "Residual from Poisson Model" or "P-value from Fisher's Exact Test".

data_inputs

Description of one or more data inputs needed to create the dfInput data.frame that is passed to the assessment function.

data_filters

Description of parameters to use to filter when creating input data (dfInput)

Source

https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html