Identity Analysis.

Used in the data pipeline between Transform and Flag to rename KRI and Score columns.

More information can be found in The Identity Method of the KRI Method vignette.

Usage

Analyze_Identity(dfTransformed, strValueCol = "Metric")

Arguments

dfTransformed

data.frame Transformed data for analysis. Data should have one record per site with expected columns: GroupID, GroupLevel, Numerator, Denominator, and Metric. For more details see the Data Model vignette: vignette("DataModel", package = "gsm"). For this function, dfTransformed should typically be created using Transform_Count().

strValueCol

character Name of column that will be copied as Score

Value

data.frame with one row per site with columns: GroupID, TotalCount, Metric, and Score.

Examples

dfTransformed <- Transform_Count(analyticsInput, strCountCol = "Numerator")
dfAnalyzed <- Analyze_Identity(dfTransformed)
#> `Score` column created from `Metric`.