QTL Analysis for Binary and Rate Outcomes.

Analyze_QTL(dfTransformed, nConfLevel = 0.95, strOutcome = "binary")

Arguments

dfTransformed

data.frame Transformed data for analysis. Data should have one record per site with expected columns: GroupID, GroupLevel, Numerator, Denominator, and Metric. For more details see the Data Pipeline vignette: vignette("DataPipeline", package = "gsm"). For this function, dfTransformed should typically be created using Transform_Rate().

nConfLevel

numeric specified confidence interval for QTL analysis.

strOutcome

character indicates statistical test used for QTL analysis. One of rate or binary.

Value

data.frame with one row with columns: GroupID, Numerator, Denominator, Metric, Method, ConfLevel, Estimate, LowCI, UpCI and Score.

Details

Creates confidence intervals for the observed proportion of participants for the event of interest. Uses the exact binomial (stats::binom.test) or poisson test (stats::poisson.test).

Statistical Methods

TODO Coming soon ...

Examples

# binary
dfTransformed <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(122),
  Denominator  = c(1301),
  Metric       = c(0.0938)
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "binary")

# rate
dfTransformedRate <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(4473),
  Denominator  = c(6.19),
  Metric       = c(723)
)
dfAnalyzed <- Analyze_QTL(dfTransformedRate, strOutcome = "rate")