Evaluates sites where subject consent was:

not given

never obtained

not followed by subject randomization

obtained after subject randomization

Consent_Assess(
  dfInput,
  nThreshold = 0.5,
  lMapping = yaml::read_yaml(system.file("mappings", "Consent_Assess.yaml", package =
    "gsm")),
  strGroup = "Site",
  nMinDenominator = NULL,
  bQuiet = TRUE
)

Arguments

dfInput

data.frame Input data, a data frame with one record per subject.

nThreshold

numeric Threshold specification. Default: 0.5

lMapping

list Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined Consent Assessment mapping.

strGroup

character Grouping variable. "Site" (the default) uses the column named in mapping$strSiteCol. Other valid options using the default mapping are "Study", "Country", and "CustomGroup".

nMinDenominator

numeric Specifies the minimum denominator required to return a score and calculate a flag. Default: NULL

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list

lData, a named list with:

each data frame in the data pipeline

dfTransformed, returned by Transform_Count()

dfAnalyzed, returned by Analyze_Identity()

dfFlagged, returned by Flag()

dfSummary, returned by Summarize()

list lCharts, a named list with:

barMetric, a ggplot2 object returned by Visualize_Score() using strType == "metric"

barScore, a ggplot2 object returned by Visualize_Score() using strType == "score"

list lChecks, a named list with:

dfInput, a named list returned by is_mapping_valid()

status, a boolean returned by is_mapping_valid()

mapping, a named list that is provided as an argument to the lMapping parameter in Consent_Assess()

spec, a named list used to define variable specifications

Details

Consent Assessment

The Consent Assessment uses the standard GSM data pipeline to flag sites with consent issues. Additional details regarding the data pipeline and statistical methods are described below.

Data specification

Domain	Column Key	Default Value	Required?	Require Unique Values?

dfInput	strIDCol	SubjectID	TRUE	TRUE

dfInput	strCountCol	Count	TRUE	FALSE

dfInput	strGroupCol	SiteID	TRUE	FALSE

Examples

dfInput <- Consent_Map_Raw()
consent_assessment <- Consent_Assess(dfInput)