Load assessments from a package/directory

Load assessments from a package/directory

MakeAssessmentList(path = "assessments", package = "gsm")

Arguments

path

character The location of assessment YAML files. If package is specified, function will look in /inst folder.

package

character package with assessments

Value

list A list of assessments with workflow and parameter metadata.

Details

MakeAssessmentList() is a utility function that creates a workflow mapping for assessments used in Study_Assess().

Examples

MakeAssessmentList(path = "assessments", package = "gsm")
#> $ae
#> $ae$label
#> [1] "Treatment-Emergent Adverse Events"
#> 
#> $ae$tags
#> $ae$tags$Assessment
#> [1] "Safety"
#> 
#> $ae$tags$Label
#> [1] "AEs"
#> 
#> 
#> $ae$workflow
#> $ae$workflow[[1]]
#> $ae$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $ae$workflow[[1]]$inputs
#> [1] "dfAE"
#> 
#> $ae$workflow[[1]]$output
#> [1] "dfAE"
#> 
#> $ae$workflow[[1]]$params
#> $ae$workflow[[1]]$params$strDomain
#> [1] "dfAE"
#> 
#> $ae$workflow[[1]]$params$strColParam
#> [1] "strTreatmentEmergentCol"
#> 
#> $ae$workflow[[1]]$params$strValParam
#> [1] "strTreatmentEmergentVal"
#> 
#> 
#> 
#> $ae$workflow[[2]]
#> $ae$workflow[[2]]$name
#> [1] "AE_Map_Raw"
#> 
#> $ae$workflow[[2]]$inputs
#> [1] "dfAE"   "dfSUBJ"
#> 
#> $ae$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> 
#> $ae$workflow[[3]]
#> $ae$workflow[[3]]$name
#> [1] "AE_Assess"
#> 
#> $ae$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $ae$workflow[[3]]$output
#> [1] "lResults"
#> 
#> $ae$workflow[[3]]$params
#> $ae$workflow[[3]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $ae$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/ae.yaml"
#> 
#> $ae$name
#> [1] "ae"
#> 
#> 
#> $consent
#> $consent$label
#> [1] "Consent Workflow"
#> 
#> $consent$tags
#> $consent$tags$Assessment
#> [1] "Consent"
#> 
#> $consent$tags$Label
#> [1] "Consent"
#> 
#> 
#> $consent$workflow
#> $consent$workflow[[1]]
#> $consent$workflow[[1]]$name
#> [1] "Consent_Map_Raw"
#> 
#> $consent$workflow[[1]]$inputs
#> [1] "dfCONSENT" "dfSUBJ"   
#> 
#> $consent$workflow[[1]]$output
#> [1] "dfInput"
#> 
#> 
#> $consent$workflow[[2]]
#> $consent$workflow[[2]]$name
#> [1] "Consent_Assess"
#> 
#> $consent$workflow[[2]]$inputs
#> [1] "dfInput"
#> 
#> $consent$workflow[[2]]$output
#> [1] "lResults"
#> 
#> 
#> 
#> $consent$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/consent.yaml"
#> 
#> $consent$name
#> [1] "consent"
#> 
#> 
#> $ie
#> $ie$label
#> [1] "Inclusion/Exclusion (IE) Workflow"
#> 
#> $ie$tags
#> $ie$tags$Assessment
#> [1] "IE"
#> 
#> $ie$tags$Label
#> [1] "IE"
#> 
#> 
#> $ie$workflow
#> $ie$workflow[[1]]
#> $ie$workflow[[1]]$name
#> [1] "IE_Map_Raw"
#> 
#> $ie$workflow[[1]]$inputs
#> [1] "dfIE"   "dfSUBJ"
#> 
#> $ie$workflow[[1]]$output
#> [1] "dfInput"
#> 
#> 
#> $ie$workflow[[2]]
#> $ie$workflow[[2]]$name
#> [1] "IE_Assess"
#> 
#> $ie$workflow[[2]]$inputs
#> [1] "dfInput"
#> 
#> $ie$workflow[[2]]$output
#> [1] "lResults"
#> 
#> 
#> 
#> $ie$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/ie.yaml"
#> 
#> $ie$name
#> [1] "ie"
#> 
#> 
#> $importantpd
#> $importantpd$label
#> [1] "Protocol Deviation (PD) Workflow - Important"
#> 
#> $importantpd$tags
#> $importantpd$tags$Assessment
#> [1] "PD"
#> 
#> $importantpd$tags$Label
#> [1] "Important PD"
#> 
#> 
#> $importantpd$workflow
#> $importantpd$workflow[[1]]
#> $importantpd$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $importantpd$workflow[[1]]$inputs
#> [1] "dfPD"
#> 
#> $importantpd$workflow[[1]]$output
#> [1] "dfPD"
#> 
#> $importantpd$workflow[[1]]$params
#> $importantpd$workflow[[1]]$params$strDomain
#> [1] "dfPD"
#> 
#> $importantpd$workflow[[1]]$params$strColParam
#> [1] "strImportantCol"
#> 
#> $importantpd$workflow[[1]]$params$strValParam
#> [1] "strImportantVal"
#> 
#> 
#> 
#> $importantpd$workflow[[2]]
#> $importantpd$workflow[[2]]$name
#> [1] "PD_Map_Raw"
#> 
#> $importantpd$workflow[[2]]$inputs
#> [1] "dfPD"   "dfSUBJ"
#> 
#> $importantpd$workflow[[2]]$output
#> [1] "dfInput"
#> 
#> 
#> $importantpd$workflow[[3]]
#> $importantpd$workflow[[3]]$name
#> [1] "PD_Assess"
#> 
#> $importantpd$workflow[[3]]$inputs
#> [1] "dfInput"
#> 
#> $importantpd$workflow[[3]]$output
#> [1] "lResults"
#> 
#> 
#> 
#> $importantpd$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/importantpd.yaml"
#> 
#> $importantpd$name
#> [1] "importantpd"
#> 
#> 
#> $pd
#> $pd$label
#> [1] "Protocol Deviation (PD) Workflow"
#> 
#> $pd$tags
#> $pd$tags$Assessment
#> [1] "PD"
#> 
#> $pd$tags$Label
#> [1] "PD"
#> 
#> 
#> $pd$workflow
#> $pd$workflow[[1]]
#> $pd$workflow[[1]]$name
#> [1] "PD_Map_Raw"
#> 
#> $pd$workflow[[1]]$inputs
#> [1] "dfPD"   "dfSUBJ"
#> 
#> $pd$workflow[[1]]$output
#> [1] "dfInput"
#> 
#> 
#> $pd$workflow[[2]]
#> $pd$workflow[[2]]$name
#> [1] "PD_Assess"
#> 
#> $pd$workflow[[2]]$inputs
#> [1] "dfInput"
#> 
#> $pd$workflow[[2]]$output
#> [1] "lResults"
#> 
#> 
#> 
#> $pd$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/pd.yaml"
#> 
#> $pd$name
#> [1] "pd"
#> 
#> 
#> $sae
#> $sae$label
#> [1] "Treatment-Emergent Serious Adverse Events by SOC"
#> 
#> $sae$tags
#> $sae$tags$Assessment
#> [1] "Safety"
#> 
#> $sae$tags$Label
#> [1] "AEs Serious"
#> 
#> 
#> $sae$workflow
#> $sae$workflow[[1]]
#> $sae$workflow[[1]]$name
#> [1] "FilterDomain"
#> 
#> $sae$workflow[[1]]$inputs
#> [1] "dfAE"
#> 
#> $sae$workflow[[1]]$output
#> [1] "dfAE"
#> 
#> $sae$workflow[[1]]$params
#> $sae$workflow[[1]]$params$strDomain
#> [1] "dfAE"
#> 
#> $sae$workflow[[1]]$params$strColParam
#> [1] "strTreatmentEmergentCol"
#> 
#> $sae$workflow[[1]]$params$strValParam
#> [1] "strTreatmentEmergentVal"
#> 
#> 
#> 
#> $sae$workflow[[2]]
#> $sae$workflow[[2]]$name
#> [1] "FilterDomain"
#> 
#> $sae$workflow[[2]]$inputs
#> [1] "dfAE"
#> 
#> $sae$workflow[[2]]$output
#> [1] "dfAE"
#> 
#> $sae$workflow[[2]]$params
#> $sae$workflow[[2]]$params$strDomain
#> [1] "dfAE"
#> 
#> $sae$workflow[[2]]$params$strColParam
#> [1] "strSeriousCol"
#> 
#> $sae$workflow[[2]]$params$strValParam
#> [1] "strSeriousVal"
#> 
#> 
#> 
#> $sae$workflow[[3]]
#> $sae$workflow[[3]]$name
#> [1] "AE_Map_Raw"
#> 
#> $sae$workflow[[3]]$inputs
#> [1] "dfAE"   "dfSUBJ"
#> 
#> $sae$workflow[[3]]$output
#> [1] "dfInput"
#> 
#> 
#> $sae$workflow[[4]]
#> $sae$workflow[[4]]$name
#> [1] "AE_Assess"
#> 
#> $sae$workflow[[4]]$inputs
#> [1] "dfInput"
#> 
#> $sae$workflow[[4]]$output
#> [1] "lResults"
#> 
#> $sae$workflow[[4]]$params
#> $sae$workflow[[4]]$params$strMethod
#> [1] "poisson"
#> 
#> 
#> 
#> 
#> $sae$path
#> [1] "/home/runner/work/_temp/Library/gsm/assessments/sae.yaml"
#> 
#> $sae$name
#> [1] "sae"
#> 
#>