Generate a Summary data.frame for use in reports

Generate a summary table for a report by joining the provided results data frame with the site-level metadata from dfGroups, and filter and arrange the data based on provided conditions.

Usage

MakeMetricTable(
  dfResults,
  dfGroups = NULL,
  strGroupLevel = c("Site", "Country", "Study"),
  strGroupDetailsParams = NULL,
  vFlags = c(-2, -1, 1, 2)
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate. For this function, dfResults must be filtered to a single KRI (MetricID).

dfGroups

data.frame Group-level metadata dictionary. Created by passing CTMS site and study data to MakeLongMeta(). Expected columns: GroupID, GroupLevel, Param, Value.

strGroupLevel

group level for the table

strGroupDetailsParams

one or more parameters from dfGroups to be added as columns in the table

vFlags

integer List of flag values to include in output table. Default: c(-2, -1, 1, 2).

Value

A data.frame containing the summary table

Examples

# site-level report
MakeMetricTable(
  dfResults = reportingResults %>%
    dplyr::filter(.data$MetricID == "kri0001") %>%
    FilterByLatestSnapshotDate(),
  dfGroups = reportingGroups
)
#>                         Group Enrolled Numerator Denominator Metric Score Flag
#> 1          0X139 (el-Khalifa)        2        67        1863   0.04 -2.96   -2
#> 2  0X186 (Gallegos-Neiberger)        2        98        2518   0.04 -2.69   -2
#> 3               0X162 (Campa)        2       186        4433   0.04 -2.54   -2
#> 4               0X113 (Wilde)        3       359        8172   0.04 -2.53   -2
#> 5               0X025 (Smith)        2       136        3210   0.04 -2.04   -2
#> 6             0X155 (al-Sani)        3       379        6887   0.06  2.37    1
#> 7             0X119 (Cousins)        3       260        4695   0.06  2.08    1
#> 8              0X091 (Ostrom)        2       195        4462   0.04 -1.95   -1
#> 9               0X138 (Black)        2        58        1454   0.04 -1.85   -1
#> 10             0X020 (Hamlin)        3       268        5985   0.04 -1.83   -1
#> 11          0X092 (al-Badour)        3       213        4795   0.04 -1.77   -1
#> 12     0X121 (Rodriguez-Cruz)        2       111        2569   0.04 -1.61   -1
#> 13              0X148 (Suess)        2       167        3742   0.04 -1.50   -1
#> 14           0X173 (el-Yusuf)        2       200        4436   0.05 -1.47   -1
#> 15           0X013 (el-Allam)        1        96        2215   0.04 -1.46   -1
#> 16              0X164 (Jiron)        1        96        2215   0.04 -1.46   -1
#> 17           0X143 (el-Salih)        3       216        4759   0.05 -1.42   -1
#> 18           0X046 (al-Daoud)        5       432        9283   0.05 -1.42   -1
#> 19          0X016 (el-Shaker)        3       245        5366   0.05 -1.41   -1
#> 20            0X176 (Salazar)        3       285        6191   0.05 -1.36   -1
#> 21            0X105 (Patalan)        2       170        3770   0.05 -1.36   -1
#> 22              0X134 (Keese)        2       143        3165   0.05 -1.22   -1