Create Labels for Parameters

Convert a vector of parameters to labels in Title Case. MakeParamLabels adds a Labels column to a data.frame that has a Params column (such as dfGroups), while MakeParamLabelsList returns just the list of named parameters.

Usage

MakeParamLabels(dfGroups, lParamLabels = NULL)

MakeParamLabelsList(chrParams, lParamLabels = NULL)

Arguments

dfGroups

data.frame Group-level metadata dictionary. Created by passing CTMS site and study data to MakeLongMeta(). Expected columns: GroupID, GroupLevel, Param, Value.

lParamLabels

list Labels for parameters, with the parameters as names, and the label as value.

chrParams

A character vector of parameters, or a list that can be coerced to a character vector.

Value

dfGroups with an added Label column, or a list of labeled parameters.

Examples

head(gsm::reportingGroups)
#>   GroupID            Param Value GroupLevel
#> 1      10 ParticipantCount    20       Site
#> 2      10        SiteCount     1       Site
#> 3     100 ParticipantCount     2       Site
#> 4     100        SiteCount     1       Site
#> 5     101 ParticipantCount     3       Site
#> 6     101        SiteCount     1       Site
MakeParamLabels(head(gsm::reportingGroups))
#>   GroupID            Param Value GroupLevel             Label
#> 1      10 ParticipantCount    20       Site Participant Count
#> 2      10        SiteCount     1       Site        Site Count
#> 3     100 ParticipantCount     2       Site Participant Count
#> 4     100        SiteCount     1       Site        Site Count
#> 5     101 ParticipantCount     3       Site Participant Count
#> 6     101        SiteCount     1       Site        Site Count
MakeParamLabels(
  head(gsm::reportingGroups),
  list(ParticipantCount = "Number of Participants")
)
#>   GroupID            Param Value GroupLevel                  Label
#> 1      10 ParticipantCount    20       Site Number of Participants
#> 2      10        SiteCount     1       Site             Site Count
#> 3     100 ParticipantCount     2       Site Number of Participants
#> 4     100        SiteCount     1       Site             Site Count
#> 5     101 ParticipantCount     3       Site Number of Participants
#> 6     101        SiteCount     1       Site             Site Count
MakeParamLabelsList(head(gsm::reportingGroups$Param))
#> $ParticipantCount
#> [1] "Participant Count"
#> 
#> $SiteCount
#> [1] "Site Count"
#> 
#> $ParticipantCount
#> [1] "ParticipantCount"
#> 
#> $SiteCount
#> [1] "SiteCount"
#> 
#> $ParticipantCount
#> [1] "ParticipantCount"
#> 
#> $SiteCount
#> [1] "SiteCount"
#>