Make_Snapshot() ingests data from a variety of sources, and runs KRIs and/or QTLs based on the list provided in lAssessments.

For more context about the inputs and outputs of Make_Snapshot(), refer to the GSM Data Pipeline Vignette, specifically Appendix 2 - Data Model Specifications

Make_Snapshot(
  lMeta = list(config_param = clindata::config_param, config_workflow =
    clindata::config_workflow, meta_params = gsm::meta_param, meta_site =
    clindata::ctms_site, meta_study = clindata::ctms_study, meta_workflow =
    gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfPD =
    clindata::ctms_protdev, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp %>%
    filter(.data$phase == "Blinded Study Drug Completion"), dfDATACHG =
    clindata::edc_data_points, dfDATAENT = clindata::edc_data_pages, dfQUERY =
    clindata::edc_queries, dfENROLL = clindata::rawplus_enroll),
  lMapping = c(yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_ctms.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_edc.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_adam.yaml", package =
    "gsm"))),
  lAssessments = NULL,
  strAnalysisDate = NULL,
  bUpdateParams = FALSE,
  cPath = NULL,
  bQuiet = TRUE,
  bFlowchart = FALSE
)

Arguments

lMeta

list a named list of data frames containing metadata, configuration, and workflow parameters for a given study. See the Data Model Vignette - Appendix 2 - Data Model Specifications for detailed specifications.

lData

list a named list of domain-level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lMapping

list Column metadata with structure domain$key, where key contains the name of the column. Default: package-defined mapping for raw+.

lAssessments

list a named list of metadata defining how each assessment should be run. By default, MakeWorkflowList() imports YAML specifications from inst/workflow.

strAnalysisDate

character date that the data was pulled/wrangled/snapshot. Note: date should be provided in format: YYYY-MM-DD.

bUpdateParams

logical if TRUE, configurable parameters found in lMeta$config_param will overwrite the default values in lMeta$meta_params. Default: FALSE.

cPath

character a character string indicating a working directory to save .csv files; the output of the snapshot.

bQuiet

logical Suppress warning messages? Default: TRUE

bFlowchart

logical Create flowchart to show data pipeline? Default: FALSE

Value

list

lSnapshot, a named list with a data.frame for each component of the gsm data model.

status_study

status_site

status_workflow

status_param

results_summary

results_analysis

results_bounds

meta_workflow

meta_param

Details

Make Snapshot - create and export data model.

Return - Data Model Description

A named list of data.frames:

status_study, with columns:

studyid - Unique Study ID

enrolled_sites - # of enrolled sites

enrolled_participants - # of enrolled participants

planned_sites - # of planned sites

planned_participants - # of planned participants

title - Protocol title

nickname - Protocol nickname

enrolled_sites_ctms - # of enrolled sites (ctms)

enrolled_participants_ctms - # of enrolled participants (ctms)

fpfv - First-patient first visit date

lpfv - Last-patient first visit date

lplv - Last-patient last visit date

ta - Therapeutic Area

indication - Indication

phase - Phase

status - Study Status

rbm_flag - Risk-based monitoring flag

product - Product

protocol_type - Protocol type

protocol_row_id - Protocol row ID

est_fpfv - Estimated first-patient first visit date

est_lpfv - Estimated last-patient first visit date

est_lplv - Estimated last-patient last visit date

protocol_product_number - Protocol product number

gsm_analysis_date - Date that Make_Snapshot was run

status_site, with columns:

studyid - Unique Study ID

siteid - Unique Site ID

institution - Institution Name

status - Site Status

enrolled_participants - # of enrolled participants

start_date - Site Activation Date

city - City

state - State

country - Country

invname - Investigator name

protocol_row_id - Protocol Row ID

site_num - Site number

site_row_id - Site Row ID

pi_number - Principal Investigator Number

pi_last_name- Principal Investigator Last Name

pi_first_name - Principal Investigator First Name

is_satellite - Is location a satellite campus/site?

gsm_analysis_date - Date that Make_Snapshot was run

status_workflow, with columns:

studyid - Unique Study ID

workflowid - Unique Workflow ID

gsm_version - GSM version of workflow

active - Is workflow active?

status - Did workflow run?

notes - Status notes

gsm_analysis_date - Date that Make_Snapshot was run

status_param, with columns:

study_id - Unique Study ID

workflowid - Unique Workflow ID

gsm_version - GSM version of workflow

param - Parameter name

index - Index value for parameter

value - Value for parameter at index

gsm_analysis_date - Date that Make_Snapshot was run

results_summary, with columns:

studyid - Unique Study ID

workflowid - Unique Workflow ID

groupid - Unique Group ID (e.g. SiteID for KRI workflows)

numerator - Metric numerator

denominator - Metric denominator

metric - Metric value

score - Statistical Score

flag - Flag

gsm_analysis_date - Date that Make_Snapshot was run

results_bounds, with columns:

studyid - Unique Study ID

workflowid - Unique Workflow ID

threshold - Threshold

numerator - Y value

denominator - X value

log_denominator - Log X value

gsm_analysis_date - Date that Make_Snapshot was run

meta_workflow, with columns:

workflowid - Unique Workflow ID

gsm_version - GSM Version of workflow

group - Group for workflow - Site, Country, Region, or Study

abbreviation - Abbreviation for KRI/assessment

metric - Metric Label

numerator - Numerator Label

denominator - Denominator Label

outcome - Metric Outcome Type (e.g. rate)

model - Statistical Model used to create score

score - Score Label

data_inputs - Data inputs for workflow

data_filters - Filters applied in workflow

gsm_analysis_date - Date that Make_Snapshot was run

meta_param, with columns:

workflowid - Unique Workflow ID

gsm_version - GSM version of workflow

param - Parameter name

index - Parameter index

default - Default value for parameter at index

configurable - Is parameter configurable?

gsm_analysis_date - Date that Make_Snapshot was run

Examples

# run with default testing data
if (FALSE) {
snapshot <- Make_Snapshot()
}