Create interactive timeline for study data

Usage

Make_Timeline(
  status_study,
  longitudinal = NULL,
  n_breaks = 10,
  date_format = "%b\n%Y",
  bInteractive = TRUE
)

Arguments

status_study

the table containing study data of interest

longitudinal

optional argument for supplying longitudinal snapshot information

n_breaks

the number of breaks to include in the x-axis. NOTE pretty breaks are used to assign break locations and may give more or less that the number of breaks specified.

date_format

The format to apply to the x-axis labels

bInteractive

logical argument to make plot interactive or not.

Value

plot interactive timeline plot.

Examples

if (require("gggenes") && require("ggiraph")) {
  Make_Timeline(status_study = clindata::ctms_study)
}
#> Loading required package: gggenes
#> Loading required package: ggiraph