Merge Domain data with subject-level data

MergeSubjects is a helper function used within mapping functions to join subject-level data with domain-level data, with the ability to impute zeros when necessary.

MergeSubjects(
  dfDomain,
  dfSUBJ,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  vRemoval = NULL,
  bQuiet = TRUE
)

Arguments

dfDomain

data.frame Subject-level domain data containing one record per participant.

dfSUBJ

data.frame Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)

strIDCol

character Vector of length 1. The name of ID Column - Default: SubjectID

vFillZero

vector Column names from dfDomain to fill with zeros when no matching row is found for an ID in dfSUBJ.

vRemoval

vector Column names from dfDomain or dfSUBJ to check for zeros or NA for rows to be removed for non-meaningful records.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one record per strIDCol.

Examples

MergeSubjects(
  dfDomain = clindata::rawplus_consent,
  dfSUBJ = clindata::rawplus_dm,
  strIDCol = "subjid"
)
#> # A tibble: 1,301 × 50
#>    studyid    siteid invid scrnid subjid subje…¹ datap…² datap…³ folde…⁴ insta…⁵
#>    <chr>      <chr>  <chr> <chr>  <chr>  <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 AA-AA-000… 5      0X167 113    0496   X16704… Demogr… 1       Screen… Screen…
#>  2 AA-AA-000… 78     0X002 142    1350   X00213… Demogr… 1       Screen… Screen…
#>  3 AA-AA-000… 139    0X052 112    0539   X05205… Demogr… 1       Screen… Screen…
#>  4 AA-AA-000… 162    0X049 142    0329   X04903… Demogr… 1       Screen… Screen…
#>  5 AA-AA-000… 29     0X116 069    0429   X11604… Demogr… 1       Screen… Screen…
#>  6 AA-AA-000… 143    0X153 073    1218   X15312… Demogr… 1       Screen… Screen…
#>  7 AA-AA-000… 173    0X124 145    0808   X12408… Demogr… 1       Screen… Screen…
#>  8 AA-AA-000… 189    0X093 087    1314   X09313… Demogr… 1       Screen… Screen…
#>  9 AA-AA-000… 58     0X091 020    1236   X09112… Demogr… 1       Screen… Screen…
#> 10 AA-AA-000… 167    0X059 061    0163   X05901… Demogr… 1       Screen… Screen…
#> # … with 1,291 more rows, 40 more variables: recordid <chr>, record_dt <chr>,
#> #   recordposition <dbl>, mincreated_dts <chr>, maxupdated_dts <chr>,
#> #   brthdat <chr>, agerep <chr>, sex <chr>, ethnic <chr>, race <chr>,
#> #   raceoth <chr>, racescat <chr>, racesoth <chr>, subjid_nsv <chr>,
#> #   scrnid_nsv <chr>, subjinit_nsv <chr>, invid_nsv <chr>, subject_nsv <chr>,
#> #   instanceid_nsv <dbl>, folder_nsv <chr>, folderseq_nsv <dbl>,
#> #   brthdat_nsv <dttm>, brthdat_yy_nsv <dbl>, brthdat_mm_nsv <dbl>, …