Merge Domain data with subject-level data shell

Merge Domain data with subject-level data shell

MergeSubjects(
  dfDomain,
  dfSubjects,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  bQuiet = TRUE
)

Arguments

dfDomain

Subject-level domain data containing one record per participant.

dfSubjects

Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)

strIDCol

name of ID Column - default='SubjectID'

vFillZero

Columns from dfDomain to fill with zeros when no matching row is found in for an ID in dfSubject

bQuiet

Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.

Value

data set with one record per IDCol

Examples

MergeSubjects(dfDomain = clindata::rawplus_consent,
              dfSubjects = clindata::rawplus_subj,
              strIDCol = "SubjectID")
#> # A tibble: 1,297 × 19
#>    SubjectID ScreenID SiteID ScreenFlag RandDate   RandFlag FirstStudyDate     
#>    <chr>     <chr>    <chr>  <chr>      <date>     <chr>    <dttm>             
#>  1 1397      153      X010X  Y          2012-09-02 Y        2012-09-02 00:00:00
#>  2 1396      108      X102X  Y          2017-05-08 Y        2017-05-09 00:00:00
#>  3 1395      087      X143X  Y          2004-09-23 Y        2004-09-24 00:00:00
#>  4 1394      142      X090X  Y          2010-08-21 Y        2010-08-22 00:00:00
#>  5 1393      153      X130X  Y          2011-10-09 Y        2011-10-10 00:00:00
#>  6 1392      044      X128X  Y          2010-01-01 Y        2010-01-02 00:00:00
#>  7 1391      058      X025X  Y          2005-04-27 Y        2005-04-28 00:00:00
#>  8 1390      002      X086X  Y          2006-08-21 Y        2006-08-22 00:00:00
#>  9 1389      093      X097X  Y          2012-07-10 Y        2012-07-11 00:00:00
#> 10 1388      092      X021X  Y          2015-07-15 Y        2015-07-16 00:00:00
#> # … with 1,287 more rows, and 12 more variables: LastStudyDate <date>,
#> #   TimeOnStudy <dbl>, FirstDoseDate <dttm>, LastDoseDate <date>,
#> #   TimeOnTreatment <dbl>, StudCompletion <chr>, StudDCReason <chr>,
#> #   TrtCompletion <chr>, TrtDCReason <chr>, CONSENT_DATE <date>,
#> #   CONSENT_TYPE <chr>, CONSENT_VALUE <chr>