Merge Domain data with subject-level data.

MergeSubjects is a helper function used within mapping functions to join subject-level data with domain-level data, with the ability to impute zeros when necessary.

MergeSubjects(
  dfDomain,
  dfSUBJ,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  vRemoval = NULL,
  bQuiet = TRUE
)

Arguments

dfDomain

data.frame Subject-level domain data containing one record per participant.

dfSUBJ

data.frame Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)

strIDCol

character Vector of length 1. The name of ID Column - Default: SubjectID

vFillZero

vector Column names from dfDomain to fill with zeros when no matching row is found for an ID in dfSUBJ.

vRemoval

vector Column names from dfDomain or dfSUBJ to check for zeros or NA for rows to be removed for non-meaningful records.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

data.frame with one record per ID variable specified in the strIDCol parameter.

Details

Examples

MergeSubjects(
  dfDomain = clindata::rawplus_consent,
  dfSUBJ = clindata::rawplus_dm,
  strIDCol = "subjid"
)
#> # A tibble: 1,301 × 51
#>    studyid        siteid invid scrnid subjid subjectid   datapagename datapageid
#>    <chr>          <chr>  <chr> <chr>  <chr>  <chr>       <chr>        <chr>     
#>  1 AA-AA-000-0000 5      0X167 113    0496   X1670496-1… Demographics 1         
#>  2 AA-AA-000-0000 78     0X002 142    1350   X0021350-1… Demographics 1         
#>  3 AA-AA-000-0000 139    0X052 112    0539   X0520539-1… Demographics 1         
#>  4 AA-AA-000-0000 162    0X049 142    0329   X0490329-1… Demographics 1         
#>  5 AA-AA-000-0000 29     0X116 069    0429   X1160429-0… Demographics 1         
#>  6 AA-AA-000-0000 143    0X153 073    1218   X1531218-0… Demographics 1         
#>  7 AA-AA-000-0000 173    0X124 145    0808   X1240808-1… Demographics 1         
#>  8 AA-AA-000-0000 189    0X093 087    1314   X0931314-0… Demographics 1         
#>  9 AA-AA-000-0000 58     0X091 020    1236   X0911236-0… Demographics 1         
#> 10 AA-AA-000-0000 167    0X059 061    0163   X0590163-0… Demographics 1         
#> # ℹ 1,291 more rows
#> # ℹ 43 more variables: foldername <chr>, instancename <chr>, recordid <chr>,
#> #   record_dt <chr>, recordposition <dbl>, mincreated_dts <chr>,
#> #   maxupdated_dts <chr>, brthdat <chr>, agerep <chr>, sex <chr>, ethnic <chr>,
#> #   race <chr>, raceoth <chr>, racescat <chr>, racesoth <chr>,
#> #   subjid_nsv <chr>, scrnid_nsv <chr>, subjinit_nsv <chr>, invid_nsv <chr>,
#> #   subject_nsv <chr>, instanceid_nsv <dbl>, folder_nsv <chr>, …