Generate a summary table for a report

This function generates a summary table for a report by joining the provided results data frame with the site-level metadata from dfGroups. It then filters and arranges the data based on certain conditions and displays the result in a datatable.

Usage

Report_MetricTable(
  dfResults,
  dfGroups = NULL,
  strGroupLevel = c("Site", "Country", "Study"),
  strGroupDetailsParams = NULL,
  vFlags = c(-2, -1, 1, 2)
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate. For this function, dfResults must be filtered to a single KRI (MetricID).

dfGroups

data.frame Group-level metadata dictionary. Created by passing CTMS site and study data to MakeLongMeta(). Expected columns: GroupID, GroupLevel, Param, Value.

strGroupLevel

group level for the table

strGroupDetailsParams

one or more parameters from dfGroups to be added as columns in the table

vFlags

integer List of flag values to include in output table. Default: c(-2, -1, 1, 2).

Value

A datatable containing the summary table

Examples

# site-level report
Report_MetricTable(
  dfResults = reportingResults %>%
    dplyr::filter(.data$MetricID == "kri0001") %>%
    FilterByLatestSnapshotDate(),
  dfGroups = reportingGroups
)
#> <table class="table table-striped" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Group </th>
#>    <th style="text-align:left;"> Enrolled </th>
#>    <th style="text-align:right;"> Numerator </th>
#>    <th style="text-align:right;"> Denominator </th>
#>    <th style="text-align:right;"> Metric </th>
#>    <th style="text-align:right;"> Score </th>
#>    <th style="text-align:left;"> Flag </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> 0X139 (el-Khalifa) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 67 </td>
#>    <td style="text-align:right;"> 1863 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -2.96 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FF5859;overflow:visible;position:relative;"><path d="M246.6 470.6c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 402.7 361.4 265.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3l-160 160zm160-352l-160 160c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 210.7 361.4 73.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X186 (Gallegos-Neiberger) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 98 </td>
#>    <td style="text-align:right;"> 2518 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -2.69 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FF5859;overflow:visible;position:relative;"><path d="M246.6 470.6c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 402.7 361.4 265.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3l-160 160zm160-352l-160 160c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 210.7 361.4 73.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X162 (Campa) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 186 </td>
#>    <td style="text-align:right;"> 4433 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -2.54 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FF5859;overflow:visible;position:relative;"><path d="M246.6 470.6c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 402.7 361.4 265.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3l-160 160zm160-352l-160 160c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 210.7 361.4 73.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X113 (Wilde) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 359 </td>
#>    <td style="text-align:right;"> 8172 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -2.53 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FF5859;overflow:visible;position:relative;"><path d="M246.6 470.6c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 402.7 361.4 265.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3l-160 160zm160-352l-160 160c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 210.7 361.4 73.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X025 (Smith) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 136 </td>
#>    <td style="text-align:right;"> 3210 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -2.04 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FF5859;overflow:visible;position:relative;"><path d="M246.6 470.6c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 402.7 361.4 265.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3l-160 160zm160-352l-160 160c-12.5 12.5-32.8 12.5-45.3 0l-160-160c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L224 210.7 361.4 73.4c12.5-12.5 32.8-12.5 45.3 0s12.5 32.8 0 45.3z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X155 (al-Sani) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 379 </td>
#>    <td style="text-align:right;"> 6887 </td>
#>    <td style="text-align:right;"> 0.06 </td>
#>    <td style="text-align:right;"> 2.37 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 137.4c12.5-12.5 32.8-12.5 45.3 0l160 160c12.5 12.5 12.5 32.8 0 45.3s-32.8 12.5-45.3 0L224 205.3 86.6 342.6c-12.5 12.5-32.8 12.5-45.3 0s-12.5-32.8 0-45.3l160-160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X119 (Cousins) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 260 </td>
#>    <td style="text-align:right;"> 4695 </td>
#>    <td style="text-align:right;"> 0.06 </td>
#>    <td style="text-align:right;"> 2.08 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 137.4c12.5-12.5 32.8-12.5 45.3 0l160 160c12.5 12.5 12.5 32.8 0 45.3s-32.8 12.5-45.3 0L224 205.3 86.6 342.6c-12.5 12.5-32.8 12.5-45.3 0s-12.5-32.8 0-45.3l160-160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X091 (Ostrom) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 195 </td>
#>    <td style="text-align:right;"> 4462 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.95 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X138 (Black) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 58 </td>
#>    <td style="text-align:right;"> 1454 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.85 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X020 (Hamlin) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 268 </td>
#>    <td style="text-align:right;"> 5985 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.83 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X092 (al-Badour) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 213 </td>
#>    <td style="text-align:right;"> 4795 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.77 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X121 (Rodriguez-Cruz) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 111 </td>
#>    <td style="text-align:right;"> 2569 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.61 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X148 (Suess) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 167 </td>
#>    <td style="text-align:right;"> 3742 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.50 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X173 (el-Yusuf) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 200 </td>
#>    <td style="text-align:right;"> 4436 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.47 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X013 (el-Allam) </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:right;"> 96 </td>
#>    <td style="text-align:right;"> 2215 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.46 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X164 (Jiron) </td>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:right;"> 96 </td>
#>    <td style="text-align:right;"> 2215 </td>
#>    <td style="text-align:right;"> 0.04 </td>
#>    <td style="text-align:right;"> -1.46 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X143 (el-Salih) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 216 </td>
#>    <td style="text-align:right;"> 4759 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.42 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X046 (al-Daoud) </td>
#>    <td style="text-align:left;"> 5 </td>
#>    <td style="text-align:right;"> 432 </td>
#>    <td style="text-align:right;"> 9283 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.42 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X016 (el-Shaker) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 245 </td>
#>    <td style="text-align:right;"> 5366 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.41 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X176 (Salazar) </td>
#>    <td style="text-align:left;"> 3 </td>
#>    <td style="text-align:right;"> 285 </td>
#>    <td style="text-align:right;"> 6191 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.36 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X105 (Patalan) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 170 </td>
#>    <td style="text-align:right;"> 3770 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.36 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 0X134 (Keese) </td>
#>    <td style="text-align:left;"> 2 </td>
#>    <td style="text-align:right;"> 143 </td>
#>    <td style="text-align:right;"> 3165 </td>
#>    <td style="text-align:right;"> 0.05 </td>
#>    <td style="text-align:right;"> -1.22 </td>
#>    <td style="text-align:left;"> <svg aria-hidden="true" role="img" viewbox="0 0 448 512" style="height:1em;width:0.88em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:#FEAA02;overflow:visible;position:relative;"><path d="M201.4 342.6c12.5 12.5 32.8 12.5 45.3 0l160-160c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 274.7 86.6 137.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l160 160z"></path></svg> </td>
#>   </tr>
#> </tbody>
#> </table>