Run a single step in an assessment

Calls a step in an assessment workflow. Currently supports *_Map_*, *_Assess and *_FilterDomain

RunStep(lStep, lMapping, lData, lTags, bQuiet)

Arguments

lStep

single workflow step (typically defined in lAssessment$workflow). Should include the name of the function to run (lStep$name), data inputs (lStep$inputs), name of output (lStep$output) and configurable parameters (lStep$params) (if any)

lMapping

List containing expected columns in each data set.

lData

a named list of domain level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lTags

tags

bQuiet

Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.

Value

A list containing the results of the lStep$name function call should contain .$checks parameter with results from is_mapping_vald for each domain in lStep$inputs.

Examples

lStep <- MakeAssessmentList()[["ae"]][["workflow"]][[1]]

lMapping <- clindata::mapping_rawplus

lData <- list(
  dfSUBJ = clindata::rawplus_subj,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_pd,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)

lTags <- list(Study = "myStudy")

ae_step <- RunStep(lStep = lStep, lMapping = lMapping, lData = lData, lTags = lTags, bQuiet = FALSE)
#> Preparing parameters for `FilterDomain()` ...
#> Calling `FilterDomain()` ...
#> 
#> ── Checking Input Data for `FilterDomain()` ──
#> 
#> ✔ No issues found for dfAE domain
#> Filtering on AE_TE_FLAG == TRUE
#> ✔ Filtered on `AE_TE_FLAG=TRUE`, to drop 2395 rows from 4851 to 2456 rows.