Attempts to run one or more assessments (lAssessments) using shared data (lData) and metadata (lMapping). By default, the sample rawplus data from the clindata package is used, and all assessments defined in inst/workflow are evaluated. Individual assessments are run using gsm::RunAssessment()

Study_Assess(
  lData = NULL,
  lMapping = NULL,
  lAssessments = NULL,
  bQuiet = TRUE,
  bLogOutput = FALSE,
  strLogFileName = NULL
)

Arguments

lData

list A named list of domain level data frames. Names should match the values specified in lMapping and lAssessments, which are generally based on the expected inputs from X_Map_Raw.

lMapping

list A named list identifying the columns needed in each data domain.

lAssessments

list A named list of metadata defining how each assessment should be run. By default, MakeWorkflowList() imports YAML specifications from inst/workflow.

bQuiet

logical Suppress warning messages? Default: TRUE

bLogOutput

logical Send console output to log file? Default: FALSE. Note: Setting bQuiet = FALSE is recommended if logging your output.

strLogFileName

character File name for log file.

Value

list of assessments containing status information and results.

Details

Run Multiple Assessments on a Study

Examples

if (FALSE) {
results <- Study_Assess() # run using defaults
}