Make Assessment overview table

Make overview table with one row per assessment and one column per site showing flagged assessments.

Study_Table(
  dfFindings,
  bFormat = TRUE,
  bShowCounts = TRUE,
  bShowSiteScore = TRUE,
  vSiteScoreThreshold = 1,
  bColCollapse = TRUE
)

Arguments

dfFindings

dataframe containing one or more stacked findings. Findings are one record per assessment per site and have the following columns: Assessment, Label, GroupID, N, PValue, Flag. PValue is ignored in the summary table.

bFormat

Use html-friendly icons in table cells. -1 is converted to a down arrow. 1 is converted to an up arrow. 0 is not shown. Other values are left as is.

bShowCounts

Show site counts? Uses first value of N for each site given in dfFindings.

bShowSiteScore

Show a "Score" row with total number of flagged assessments for each site. TODO: add method for custom scoring in future release)

vSiteScoreThreshold

Hide sites with a site score less than this value (1 by default).

bColCollapse

Combine the Assessment and Label columns into a single "Title Column"

Value

data.frame Returns a data.frame giving assessment status (rows) by Site (column)

Examples

lData <- list(
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_pd,
  dfSUBJ = clindata::rawplus_subj
)

lAssessment <- Study_Assess(lData)

dfSummary <- lAssessment %>%
  purrr::map(~ .x$lResults) %>%
  purrr::compact() %>%
  purrr::map_df(~ .x$dfSummary)

lStudyTable <- Study_Table(dfSummary)