Bar Chart Widget

A widget that generates a bar chart of group-level metric results, plotting groups on the x-axis and the outcome (numerator, denominator, metric, or score) on the y-axis.

Usage

Widget_BarChart(
  dfResults,
  lMetric = list(),
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "Score",
  bAddGroupSelect = TRUE,
  bDebug = FALSE
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. Created by passing a list of results returned by Summarize() to BindResults(). Expected columns: GroupID, GroupLevel, Numerator, Denominator, Metric, Score, Flag, MetricID, StudyID, SnapshotDate.

lMetric

list Metric-specific metadata for use in charts and reporting. Created by passing an lWorkflow object to MakeMetric() and turing it into a list. Expected columns: File,MetricID, Group, Abbreviation, Metric, Numerator, Denominator, Model, Score, and strThreshold. For more details see the Data Model vignette: vignette("DataModel", package = "gsm").

dfGroups

data.frame Group-level metadata dictionary. Created by passing CTMS site and study data to MakeLongMeta(). Expected columns: GroupID, GroupLevel, Param, Value.

vThreshold

numeric Threshold values.

strOutcome

character Outcome variable. Default: 'Score'.

bAddGroupSelect

logical Add a dropdown to highlight sites? Default: TRUE.

bDebug

logical Print debug messages? Default: FALSE.

Details

Examples

## Filter data to one metric and snapshot
reportingResults_filter <- reportingResults %>%
  dplyr::filter(MetricID == "kri0001" & SnapshotDate == max(SnapshotDate))

reportingMetrics_filter <- reportingMetrics %>%
  dplyr::filter(MetricID == "kri0001") %>%
  as.list()

## Make chart
Widget_BarChart(
  dfResults = reportingResults_filter,
  lMetric = reportingMetrics_filter,
  vThreshold = reportingMetrics$Threshold
)