Check that a data frame contains columns and fields specified in mapping.

is_mapping_valid is a utility function that is a core feature of any assessment or workflow. This function evaluates the validity of input data based on a pre-defined, or user-defined mapping and specification.

is_mapping_valid(df, mapping, spec, bQuiet = TRUE)

Arguments

df

data.frame A data.frame to compare to mapping object.

mapping

list A named list specifying expected columns and values in df. Parameters ending in col are assumed to be column names in df, while parameters ending in val are values expected in for a corresponding column. For example, mapping=list(strSiteCol="SiteID", strSiteVal=c("001","002")) would indicate that df has a df$SiteID includes values "001" and "002".

spec

list A named list specifying parameters that should be defined in mapping, and describes how the values specified by those parameters should be used in df. Should have the following properties:

spec$vRequired - list of parameters that should be defined in mapping.

spec$vUniqueCols - list of column parameters that should not contain duplicate values

spec$vNACols - list of column parameters where NA and empty string values are acceptable.

bQuiet

logical Suppress warning messages? Default: TRUE

Value

list A list is returned with status (TRUE or FALSE), and tests_if, a list containing checks and a status and warning (if check does not pass).

Details

Examples

subj_mapping <- list(
  strIDCol = "SubjectID",
  strSiteCol = "SiteID",
  strExposureCol = "TimeOnTreatment"
)

is_mapping_valid(
  df = clindata::rawplus_dm,
  mapping = subj_mapping,
  spec = list(
    vRequired = c("strIDCol", "strSiteCol", "strExposureCol"),
    vUniqueCols = "SUBJID"
  )
)
#> $status
#> [1] FALSE
#> 
#> $tests_if
#> $tests_if$is_data_frame
#> $tests_if$is_data_frame$status
#> [1] TRUE
#> 
#> $tests_if$is_data_frame$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_required_params
#> $tests_if$has_required_params$status
#> [1] TRUE
#> 
#> $tests_if$has_required_params$warning
#> [1] NA
#> 
#> 
#> $tests_if$spec_is_list
#> $tests_if$spec_is_list$status
#> [1] TRUE
#> 
#> $tests_if$spec_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mapping_is_list
#> $tests_if$mapping_is_list$status
#> [1] TRUE
#> 
#> $tests_if$mapping_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mappings_are_character
#> $tests_if$mappings_are_character$status
#> [1] TRUE
#> 
#> $tests_if$mappings_are_character$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_expected_columns
#> $tests_if$has_expected_columns$status
#> [1] FALSE
#> 
#> $tests_if$has_expected_columns$warning
#> [1] "the following columns not found in df: SubjectID, SiteID, TimeOnTreatment"
#> 
#> 
#> $tests_if$columns_have_na
#> $tests_if$columns_have_na$status
#> [1] FALSE
#> 
#> $tests_if$columns_have_na$warning
#> [1] "NA check not run"
#> 
#> 
#> $tests_if$columns_have_empty_values
#> $tests_if$columns_have_empty_values$status
#> [1] FALSE
#> 
#> $tests_if$columns_have_empty_values$warning
#> [1] "Empty Value check not run"
#> 
#> 
#> $tests_if$cols_are_unique
#> $tests_if$cols_are_unique$status
#> [1] FALSE
#> 
#> $tests_if$cols_are_unique$warning
#> [1] "Unique Column Check not run"
#> 
#> 
#> 
#> $dim
#> [1] 1301   48
#> 

is_mapping_valid(
  df = clindata::rawplus_dm,
  mapping = subj_mapping,
  spec = list(
    vUniqueCols = "SUBJID",
    vRequiredParams = c("strIDCol", "strSiteCol", "strExposureCol", "strOtherCol")
  )
)
#> $status
#> [1] FALSE
#> 
#> $tests_if
#> $tests_if$is_data_frame
#> $tests_if$is_data_frame$status
#> [1] TRUE
#> 
#> $tests_if$is_data_frame$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_required_params
#> $tests_if$has_required_params$status
#> [1] FALSE
#> 
#> $tests_if$has_required_params$warning
#> [1] "\"mapping\" does not contain required parameters: strOtherCol"
#> 
#> 
#> $tests_if$spec_is_list
#> $tests_if$spec_is_list$status
#> [1] TRUE
#> 
#> $tests_if$spec_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mapping_is_list
#> $tests_if$mapping_is_list$status
#> [1] TRUE
#> 
#> $tests_if$mapping_is_list$warning
#> [1] NA
#> 
#> 
#> $tests_if$mappings_are_character
#> $tests_if$mappings_are_character$status
#> [1] TRUE
#> 
#> $tests_if$mappings_are_character$warning
#> [1] NA
#> 
#> 
#> $tests_if$has_expected_columns
#> $tests_if$has_expected_columns$status
#> [1] FALSE
#> 
#> $tests_if$has_expected_columns$warning
#> [1] "the following columns not found in df: SubjectID, SiteID, TimeOnTreatment"
#> 
#> 
#> $tests_if$columns_have_na
#> $tests_if$columns_have_na$status
#> [1] FALSE
#> 
#> $tests_if$columns_have_na$warning
#> [1] "NA check not run"
#> 
#> 
#> $tests_if$columns_have_empty_values
#> $tests_if$columns_have_empty_values$status
#> [1] FALSE
#> 
#> $tests_if$columns_have_empty_values$warning
#> [1] "Empty Value check not run"
#> 
#> 
#> $tests_if$cols_are_unique
#> $tests_if$cols_are_unique$status
#> [1] FALSE
#> 
#> $tests_if$cols_are_unique$warning
#> [1] "Unique Column Check not run"
#> 
#> 
#> 
#> $dim
#> [1] 1301   48
#>